Home

Julicord

Julicord.Activity

Julicord.Attachment

Julicord.Client

Julicord.DiscordChannel

Julicord.Embed

Julicord.Emoji

Julicord.Guild

Julicord.GuildMember

Julicord.Message

Julicord.MessageDelete

Julicord.Presence

Julicord.Reaction

Julicord.Role

Julicord.UnavailableGuild

Julicord.UnknownEvent

Julicord.User

Julicord.VoiceState

Julicord.Webhook

Base.open

Base.wait

Julicord.add_handler!

Julicord.clear_handlers!

Julicord.me

Julicord.state

Julicord.Client — Type.

Client(token::String) -> Client

A Discord bot.

Arguments

token::String: The bot's token.

source

Julicord.MessageDelete — Method.

A message delete event. More details here.

source

Julicord.UnknownEvent — Type.

An unknown event.

source

Julicord.User — Method.

A Discord user. More details here.

source

Julicord.add_handler! — Method.

add_handler!(c::Client, evt::Type{<:AbstractEvent}, func::Function)

Add a handler for the given event type. The handler is appended the event's current handlers.

source

Julicord.clear_handlers! — Method.

clear_handlers!(c::Client, evt::Type{<:AbstractEvent})

Removes all handlers for the given event type.

source

Julicord.me — Method.

me(c::Client) -> Dict{String, Any}

Get the client's bot user.

source

Julicord.state — Method.

state(c::Client) -> Dict{String, Any}

Get the client state.

source

Julicord.Activity — Method.

A user activity. More details here.

source

Julicord.Attachment — Method.

A message attachment. More details here.

source

Julicord.DiscordChannel — Method.

A Discord channel. More details here. Note: The name Channel is already used, hence the prefix.

source

Julicord.Embed — Method.

A message embed. More details here.

source

Julicord.Emoji — Method.

An emoji. More details here.

source

Julicord.Guild — Method.

A guild (server). More details here.

source

Julicord.GuildMember — Method.

A guild member. More details here.

source

Julicord.Message — Method.

A message. More details here.

source

Julicord.Presence — Method.

A user presence. More details here.

source

Julicord.Reaction — Method.

A reaction. More details here.

source

Julicord.Role — Method.

A role. More details here.

source

Julicord.UnavailableGuild — Method.

An unavailable guild (server). More details here.

source

Julicord.VoiceState — Method.

A voice state. More details here.

source

Julicord.Webhook — Method.

A Webhook. More details here.

source

Base.open — Method.

open(c::Client)

Log in to the Discord gateway and begin reading events.

source

Base.wait — Method.

wait(c::Client)

Wait for an open client to close.

source