Events

Events

Note that Snowflake === UInt64. Unions with Nothing indicate that a field is nullable, whereas Unions with Missing indicate that a field is optional. More details here.

Discord.AbstractEvent — Type.

An incoming event sent over the gateway. Also a catch-all event: Handlers defined on this type will execute on all events, before the specific handlers run.

source

Discord.FallbackEvent — Type.

A type for defining handlers on any events which would otherwise have no non-default handler. Handlers for this type must accept an AbstractEvent.

source

Discord.UnknownEvent — Type.

An unknown event. When an event can't be parsed, due to an unknown type or any other error, it will appear as an UnknownEvent. The fields follow the schema defined here.

source

Channels

Discord.ChannelCreate — Type.

Sent when a new DiscordChannel is created.

Fields

channel :: DiscordChannel

source

Discord.ChannelUpdate — Type.

Sent when a DiscordChannel is updated.

Fields

channel :: DiscordChannel

source

Discord.ChannelDelete — Type.

Sent when a DiscordChannel is deleted.

Fields

channel :: DiscordChannel

source

Discord.ChannelPinsUpdate — Type.

Sent when a DiscordChannel's pins are updated.

Fields

channel_id         :: UInt64
last_pin_timestamp :: Union{Nothing, DateTime}

source

Guilds

Discord.GuildCreate — Type.

Sent when a new Guild is created.

Fields

guild :: Guild

source

Discord.GuildUpdate — Type.

Sent when a Guild is updated.

Fields

guild :: Guild

source

Discord.GuildDelete — Type.

Sent when a Guild is deleted.

Fields

guild :: AbstractGuild

source

Discord.GuildBanAdd — Type.

Sent when a User is banned from a Guild.

Fields

guild_id :: UInt64
user     :: User

source

Discord.GuildBanRemove — Type.

Sent when a User is unbanned from a Guild.

Fields

guild_id :: UInt64
user     :: User

source

Discord.GuildEmojisUpdate — Type.

Sent when a Guild has its Emojis updated.

Fields

guild_id :: UInt64
emojis   :: Array{Emoji,1}

source

Discord.GuildIntegrationsUpdate — Type.

Sent when a Guild has its Integrations updated.

Fields

guild_id :: UInt64

source

Discord.GuildMemberAdd — Type.

Sent when a Member is added to a Guild.

Fields

guild_id :: UInt64
member   :: Member

source

Discord.GuildMemberRemove — Type.

Sent when a Member is removed from a Guild.

Fields

guild_id :: UInt64
user     :: User

source

Discord.GuildMemberUpdate — Type.

Sent when a Member is updated in a Guild.

Fields

guild_id :: UInt64
roles    :: Array{UInt64,1}
user     :: User
nick     :: Union{Nothing, String}

source

Discord.GuildMembersChunk — Type.

Sent when the Client requests guild members with request_guild_members.

Fields

guild_id :: UInt64
members  :: Array{Member,1}

source

Discord.GuildRoleCreate — Type.

Sent when a new Role is created in a Guild.

Fields

guild_id :: UInt64
role     :: Role

source

Discord.GuildRoleUpdate — Type.

Sent when a Role is updated in a Guild.

Fields

guild_id :: UInt64
role     :: Role

source

Discord.GuildRoleDelete — Type.

Sent when a Role is deleted from a Guild.

Fields

guild_id :: UInt64
role_id  :: UInt64

source

Messages

Discord.MessageCreate — Type.

Sent when a Message is sent.

Fields

message :: Message

source

Discord.MessageUpdate — Type.

Sent when a Message is updated.

Fields

message :: Message

source

Discord.MessageDelete — Type.

Sent when a Message is deleted.

Fields

id         :: UInt64
channel_id :: UInt64
guild_id   :: Union{Missing, UInt64}

source

Discord.MessageDeleteBulk — Type.

Sent when multiple Messages are deleted in bulk.

Fields

ids        :: Array{UInt64,1}
channel_id :: UInt64
guild_id   :: Union{Missing, UInt64}

source

Discord.MessageReactionAdd — Type.

Sent when a Reaction is added to a Message.

Fields

user_id    :: UInt64
channel_id :: UInt64
message_id :: UInt64
guild_id   :: Union{Missing, UInt64}
emoji      :: Emoji

source

Discord.MessageReactionRemove — Type.

Sent when a Reaction is removed from a Message.

Fields

user_id    :: UInt64
channel_id :: UInt64
message_id :: UInt64
guild_id   :: Union{Missing, UInt64}
emoji      :: Emoji

source

Discord.MessageReactionRemoveAll — Type.

Sent when all Reactions are removed from a Message.

Fields

channel_id :: UInt64
message_id :: UInt64
guild_id   :: Union{Missing, UInt64}

source

Presence

Discord.PresenceUpdate — Type.

Sent when a User's Presence is updated.

Fields

presence :: Presence

source

Discord.TypingStart — Type.

Sent when a User begins typing.

Fields

channel_id :: UInt64
guild_id   :: Union{Missing, UInt64}
user_id    :: UInt64
timestamp  :: Dates.DateTime

source

Discord.UserUpdate — Type.

Sent when a User's details are updated.

Fields

user :: User

source

Voice

Discord.VoiceStateUpdate — Type.

Sent when a User updates their VoiceState.

Fields

state :: VoiceState

source

Discord.VoiceServerUpdate — Type.

Sent when a Guild's voice server is updated.

Fields

token    :: String
guild_id :: UInt64
endpoint :: String

source

Webhooks

Discord.WebhooksUpdate — Type.

Sent when a DiscordChannel's Webhooks are updated.

Fields

guild_id   :: UInt64
channel_id :: UInt64

source

Connecting

Discord.Ready — Type.

Sent when the Client has authenticated, and contains the initial state.

Fields

v                :: Int64
user             :: User
private_channels :: Array{DiscordChannel,1}
guilds           :: Array{UnavailableGuild,1}
session_id       :: String
_trace           :: Array{String,1}

source

Discord.Resumed — Type.

Sent when a Client resumes its connection.

Fields

_trace :: Array{String,1}

source