Helpers

Helpers

Discord.Permission — Type.

Bitwise permission flags. More details here.

source

Discord.has_permission — Function.

has_permission(perms::Integer, perm::Permission) -> Bool

Determine whether a bitwise OR of permissions contains one Permission.

Example

julia> has_permission(0x0420, PERM_VIEW_CHANNEL)
true

julia> has_permission(0x0420, PERM_ADMINISTRATOR)
false

julia> has_permission(0x0008, PERM_MANAGE_ROLES)
true

source

Discord.permissions_in — Function.

permissions_in(m::Member, g::Guild, ch::DiscordChannel) -> Int

Compute a Member's Permissions in a DiscordChannel.

source

Discord.mention — Function.

mention(x::Union{DiscordChannel, Member, Role, User}) -> String

Get the mention string for an entity.

Examples

julia> u = User(; id=0xff);

julia> mention(u)
"<@255>"

julia> mention(Member(; user=u, nick="foo", roles=Role[], joined_at=now(), deaf=true, mute=true))
"<@!255>"

julia> mention(DiscordChannel(; id=0xff, type=CT_GUILD_TEXT))
"<#255>"

julia> mention(Role(; id=0xff, name="foo"))
"<@&255>"

source

Discord.reply — Function.

reply(
    c::Client,
    m::Message,
    content::Union{AbstractString, AbstractDict, NamedTuple, Embed};
    at::Bool=false,
) -> Future{Response}

Reply (send a message to the same DiscordChannel) to a Message. If at is set, then the message is prefixed with the sender's mention.

source

Discord.split_message — Function.

split_message(text::AbstractString) -> Vector{String}

Split a message into 2000-character chunks, preserving formatting.

Examples

```jldoctest; setup=:(using Discord) julia> split_message("foo") 1-element Array{String,1}: "foo"

julia> split_message(repeat('.', 1995) * "hello, world")[2] "hello, world"

source

Discord.plaintext — Function.

plaintext(m::Message) -> String
plaintext(c::Client, m::Message) -> String

Get the Message contents with any User mentions replaced with their plaintext. If a Client is provided, DiscordChannels Role are also replaced. However, only channels and roles stored in state are replaced; no API requests are made.

source

Discord.upload_file — Function.

upload_file(c::Client, ch::DiscordChannel, path::AbstractString; kwargs...) -> Message

Send a Message with a file Attachment. Any keywords are passed on to create_message.

source

Discord.set_game — Function.

set_game(
    c::Client,
    game::AbstractString;
    type::Union{ActivityType, Int}=AT_GAME,
    since::Union{Int, Nothing}=c.presence["since"],
    status::Union{PresenceStatus, AbstractString}=c.presence["status"],
    afk::Bool=c.presence["afk"],
    kwargs...,
) -> Bool

Shortcut for update_status to set the Client's Activity. Any additional keywords are passed into the activity section.

source

Discord.@fetch — Macro.

@fetch [functions...] block

Wrap all calls to the specified CRUD functions (create, retrieve, update, and delete) with fetch inside a block. If no functions are specified, all CRUD functions are wrapped.

Examples

Wrapping all CRUD functions:

resp = @fetch begin
    resp = create(c, Guild; name="foo")
    if resp.ok
        retrieve(c, DiscordChannel, resp.val)
    else
        error("Request failed")
    end
end

Wrapping only calls to retrieve:

future = @fetch retrieve begin
    resp = retrieve(c, DiscordChannel, 123)
    create(c, Message, resp.val; content="foo")  # Behaves normally.
end

source

Discord.@fetchval — Macro.

@fetchval [functions...] block

Wrap all calls to the specified CRUD functions (create, retrieve, update, and delete) with fetchval inside a block. If no functions are specified, all CRUD functions are wrapped.

Examples

Wrapping all CRUD functions:

channels = @fetchval begin
    guild = create(c, Guild; name="foo")
    retrieve(c, DiscordChannel, 123)
end

Wrapping only calls to retrieve:

future = @fetchval retrieve begin
    channel = retrieve(c, DiscordChannel, 123)
    create(c, Message, channel; content="foo")  # Behaves normally.
end

source