REST API

Response

Discord.Response — Type.

A wrapper around a response from the REST API. Every function which wraps a Discord REST API endpoint returns a Future which will contain a value of this type. To retrieve the Response from the Future, use fetch.

Fields

val::Union{T, Nothing}: The object contained in the HTTP response. For example, for a call to get_message, this value will be a Message.

success::Bool: The state of the request. If true, then it is safe to access val.

http_response::Union{HTTP.Messages.Response, Nothing}: The underlying HTTP response. If no HTTP request was made in the case of a cache hit, it is nothing.

Example

julia> using Discord; c = Client("token"); ch = 1234567890;

julia> fs = map(i -> send_message(c, ch, string(i)), 1:10);

julia> typeof(first(fs))
Distributed.Future

julia> typeof(fetch(first(fs)))
Discord.Response{Discord.Message}

source

Endpoints

Discord.get_message — Function.

get_message(
    c::Client,
    channel::Union{DiscordChannel, Integer},
    message::Integer,
) -> Response{Message}
get_message(c::Client, m::Message) -> Response{Message}

Get a Message from a DiscordChannel.

source

TODO