Client

Client

Discord.Client — Type.

Client(
    token::String;
    on_limit::OnLimit=LIMIT_IGNORE,
    ttl::Period=Hour(1),
    version::Int=6,
 ) -> Client

A Discord bot. Clients can connect to the gateway, respond to events, and make REST API calls to perform actions such as sending/deleting messages, kicking/banning users, etc.

To get a bot token, head here to create a new application. Once you've created a bot user, you will have access to its token.

Keywords

on_limit::OnLimit=LIMIT_IGNORE: Client's behaviour when it hits a rate limit (see "Rate Limiting" below for more details).

ttl::Period=Hour(1) Amount of time that cache entries are kept (see "Caching" below for more details).

version::Int=6: Version of the Discord API to use. Using anything but 6 is not officially supported by the Discord.jl developers.

Caching

By default, most data that comes from Discord is cached for later use. However, to avoid memory leakage, it's deleted after some time (determined by the ttl keyword). Although it's not recommended, you can also disable caching of certain data by clearing default handlers for relevant event types with clear_handlers!. For example, if you wanted to avoid caching any messages, you would clear handlers for MessageCreate and MessageUpdate events.

Rate Limiting

Discord enforces rate limits on usage of its REST API. This means you can only send so many messages in a given period, and so on. To customize the client's behaviour when encountering rate limits, use the on_limit keyword and see OnLimit.

Sharding

Sharding is handled automatically: The number of available processes is the number of shards that are created. See the sharding example for more details.

source

Base.open — Function.

open(c::Client; delay::Period=Second(7))

Connect to the Discord gateway and begin responding to events.

The delay keyword is the number of seconds between shards connecting. It can be increased from its default if you are frequently experiencing invalid sessions upon connection.

source

Base.isopen — Function.

isopen(c::Client) -> Bool

Determine whether the client is connected to the gateway.

source

Base.close — Function.

Base.close(c::Client)

Disconnect from the Discord gateway.

source

Base.wait — Function.

wait(c::Client)

Wait for an open client to close.

source

Discord.me — Function.

me(c::Client) -> Union{User, Nothing}

Get the client's bot user.

source

Event Handlers

Discord.add_handler! — Function.

add_handler!(
    c::Client,
    evt::Type{<:AbstractEvent},
    func::Function;
    tag::Symbol=gensym(),
    expiry::Union{Int, Period}=-1,
)

Add a handler for an event type. The handler should be a function which takes two arguments: A Client and an AbstractEvent (or a subtype). The handler is appended the event's current handlers.

Keywords

tag::Symbol=gensym(): A label for the handler, which can be used to remove it with delete_handler!.

expiry::Union{Int, Period}=-1: The handler's expiry. If an Int is given, the handler will run a set number of times before expiring. If a Period is given, the handler will expire after that amount of time has elapsed. The default of -1 indicates no expiry.

Note

There is no guarantee on the order in which handlers run, except that catch-all (AbstractEvent) handlers run before specific ones.

source

Discord.add_command! — Function.

add_command!(
    c::Client,
    prefix::AbstractString,
    func::Function;
    tag::Symbol=gensym(),
    expiry::Union{Int, Period}=-1,
)

Add a text command handler. The handler function should take two arguments: A Client and a Message. The keyword arguments are identical to add_handler!.

source

Discord.delete_handler! — Function.

delete_handler!(c::Client, evt::Type{<:AbstractEvent}, tag::Symbol)

Delete a single handler by event type and tag.

source

Discord.clear_handlers! — Function.

clear_handlers!(c::Client, evt::Type{<:AbstractEvent})

Remove all handlers for an event type. Using this is generally not recommended because it also clears default handlers which maintain the client state. Instead, it's preferred add handlers with specific tags and delete them with delete_handler!.

source

Gateway Commands

Discord.request_guild_members — Function.

request_guild_members(
    c::Client,
    guild_id::Union{Snowflake, Vector{Snowflake};
    query::AbstractString="",
    limit::Int=0,
) -> Bool

Request offline guild members of one or more guilds. GuildMembersChunk events are sent by the gateway in response.

More details here.

source

Discord.update_voice_state — Function.

update_voice_state(
    c::Client,
    guild_id::Snowflake,
    channel_id::Union{Snowflake, Nothing},
    self_mute::Bool,
    self_deaf::Bool,
) -> Bool

Join, move, or disconnect from a voice channel. A VoiceStateUpdate event is sent by the gateway in response.

More details here.

source

Discord.update_status — Function.

update_status(
    c::Client,
    since::Union{Int, Nothing},
    activity::Union{Activity, Nothing},
    status::PresenceStatus,
    afk::Bool,
) -> Bool

Indicate a presence or status update. A PresenceUpdate event is sent by the gateway in response.

More details here.

source