Client

Client

Discord.Client — Type.

Client(
    token::String;
    prefix::String="",
    presence::Union{Dict, NamedTuple}=Dict(),
    ttls::Dict{DataType,Union{Nothing, Period}}=Dict(),
    version::Int=6,
) -> Client

A Discord bot. Clients can connect to the gateway, respond to events, and make REST API calls to perform actions such as sending/deleting messages, kicking/banning users, etc.

Bot Token

A bot token can be acquired by creating a new application here. Make sure not to hardcode the token into your Julia code! Use an environment variable or configuration file instead.

Command Prefix

The prefix keyword specifies the command prefix, which is used by commands added with add_command!. It can be changed later, both globally and on a per-guild basis, with set_prefix!.

Presence

The presence keyword sets the bot's presence upon connection. It also sets defaults for future calls to set_game. The schema here must be followed.

Cache Control

By default, most data that comes from Discord is cached for later use. However, to avoid memory leakage, some of it is deleted after some time. The default setings are to keep everything but Messages forever, but they can be overridden with the ttls keyword. Keys can be any of the following: Guild, DiscordChannel, Message, User, Member, or Presence. Values of nothing indicate no expiry. However, the default settings are sufficient for most workloads.

If you want to entirely avoid caching certain objects, you can delete default handlers with delete_handler! and DEFAULT_HANDLER_TAG. For example, if you wanted to avoid caching any messages at all, you would delete handlers for MessageCreate and MessageUpdate events.

The cache can also be disabled/enabled permanently and temporarily with enable_cache! and disable_cache!.

API Version

The version keyword chooses the Version of the Discord API to use. Using anything but 6 is not officially supported by the Discord.jl developers.

Sharding

Sharding is handled automatically. The number of available processes is the number of shards that are created. See the sharding example for more details.

source

Discord.enable_cache! — Function.

enable_cache!(c::Client)
enable_cache!(f::Function c::Client)

Enable the cache. do syntax is also accepted.

source

Discord.disable_cache! — Function.

disable_cache!(c::Client)
disable_cache!(f::Function, c::Client)

Disable the cache. do syntax is also accepted.

source

Discord.me — Function.

me(c::Client) -> Union{User, Nothing}

Get the Client's bot user.

source

Gateway

Base.open — Function.

open(c::Client; delay::Period=Second(7))

Connect a Client to the Discord gateway.

The delay keyword is the time between shards connecting. It can be increased from its default if you are using multiple shards and frequently experiencing invalid sessions upon connection.

source

Base.isopen — Function.

isopen(c::Client) -> Bool

Determine whether the Client is connected to the gateway.

source

Base.close — Function.

close(c::Client)

Disconnect the Client from the gateway.

source

Base.wait — Function.

wait(c::Client)

Wait for an open Client to close.

source

Discord.request_guild_members — Function.

request_guild_members(
    c::Client,
    guilds::Union{Integer, Vector{<:Integer};
    query::AbstractString="",
    limit::Int=0,
) -> Bool

Request offline guild members of one or more Guilds. GuildMembersChunk events are sent by the gateway in response. More details here.

source

Discord.update_voice_state — Function.

update_voice_state(
    c::Client,
    guild::Integer,
    channel::Union{Integer, Nothing},
    mute::Bool,
    deaf::Bool,
) -> Bool

Join, move, or disconnect from a voice channel. A VoiceStateUpdate event is sent by the gateway in response. More details here.

source

Discord.update_status — Function.

update_status(
    c::Client,
    since::Union{Int, Nothing},
    activity::Union{Activity, Nothing},
    status::Union{PresenceStatus, AbstractString},
    afk::Bool,
) -> Bool

Indicate a presence or status update. A PresenceUpdate event is sent by the gateway in response. More details here.

source

Discord.heartbeat_ping — Function.

heartbeat_ping(c::Client) -> Union{Period, Nothing}

Get the Client's ping time to the gateway. If the client is not connected, or no heartbeats have been sent/acknowledged, nothing is returned.

source

Event Handlers

See Events for more details.

Discord.add_handler! — Function.

add_handler!(
    c::Client,
    T::Type{<:AbstractEvent},
    handler::Function;
    tag::Symbol=gensym(),
    predicate::Function=alwaystrue,
    fallback::Function=donothing,
    priority::Int=100,
    n::Union{Int, Nothing}=nothing,
    timeout::Union{Period, Nothing}=nothing,
    wait::Bool=false,
    compile::Bool=false,
    kwargs...,
) -> Union{Vector{Any}, Nothing}

Add an event handler. do syntax is also accepted.

Handler Function

The handler function does the real work and must take two arguments: A Client and an AbstractEvent (or a subtype).

Handler Tag

The tag keyword gives a label to the handler, which can be used to remove it with delete_handler!.

Predicate/Fallback Functions

The predicate keyword specifies a predicate function. The handler will only run if this function returns true. Otherwise, a fallback function, specified by the fallback keyword, is run. Their signatures should match that of the handler.

Handler Priority

The priority keyword indicates the handler's priority relative to other handlers for the same event. Handlers with higher values execute before those with lower ones.

Handler Expiry

Handlers can have counting and/or timed expiries. The n keyword sets the number of times a handler is run before expiring. The timeout keyword sets how long the handler remains active.

Blocking Handlers and Result Collection

To collect results from a handler, set the wait keyword along with an expiry. The call will block until the handler expires, at which point the return value of each invocation is returned in a Vector.

Forcing Precompilation

Handler functions are precompiled without running them, but it's not always successful, especially if your functions are not type-safe. If the compile keyword is set, precompilation is forced by running the predicate and handler on a randomized input. Any trailing keywords are passed to the input event constructor.

Examples

Adding a handler with a timed expiry and tag:

add_handler!(c, ChannelCreate, (c, e) -> @show e; tag=:show, timeout=Minute(1))

Adding a handler with a predicate and do syntax:

add_handler!(c, ChannelCreate; predicate=(c, e) -> length(e.channel.name) < 10) do c, e
    println(e.channel.name)
end

Aggregating results of a handler with a counting expiry:

msgs = add_handler!(c, MessageCreate, (c, e) -> e.message.content; n=5, wait=true)

Forcing precompilation and assigning a low priority:

add_handler!(c, MessageCreate, (c, e) -> @show e; priority=-10, compile=true)

source

add_handler!(
    c::Client,
    m::Module;
    tag::Symbol=gensym(),
    predicate::Function=alwaystrue,
    fallback::Function=donothing,
    n::Union{Int, Nothing}=nothing,
    timeout::Union{Period, Nothing}=nothing,
    compile::Bool=false,
)

Add all of the event handlers defined in a module. Any function you wish to use as a handler must be exported. Only functions with correct type signatures (see above) are used.

Note

If you set keywords, they are applied to all of the handlers in the module. For example, if you add two handlers for the same event type with the same tag, one of them will be immediately overwritten.

source

Discord.delete_handler! — Function.

delete_handler!(c::Client, T::Type{<:AbstractEvent}, tag::Symbol)
delete_handler!(c::Client, T::Type{<:AbstractEvent})

Delete event handlers. If no tag is supplied, all handlers for the event are deleted. Using the tagless method is generally not recommended because it also clears default handlers which maintain the client state. If you do want to delete a default handler, use DEFAULT_HANDLER_TAG.

source

Discord.DEFAULT_HANDLER_TAG — Constant.

Tag assigned to default handlers, which you can use to delete them.

source

Bot Commands

Discord.add_command! — Function.

add_command!(
    c::Client,
    name::Symbol,
    handler::Function;
    help::AbstractString="",
    parsers::Vector=[],
    separator::Union{AbstractString, AbstractChar}=' ',
    pattern::Regex=defaultpattern(name, length(parsers), separator),
    allowed::Vector{<:Integer}=Snowflake[],
    cooldown::Union{Period, Nothing}=nothing,
    fallback::Function=donothing,
    priority::Int=100,
    n::Union{Int, Nothing}=nothing,
    timeout::Union{Period, Nothing}=nothing,
    compile::Bool=false,
    kwargs...,
)

Add a text command handler. do syntax is also accepted.

Handler Function

The handler function must accept a Client and a Message. Additionally, it can accept any number of additional arguments, which are captured from pattern and parsed with parsers (see below).

Command Pattern

The pattern keyword specifies how to invoke the command. The given Regex must match the message contents after having removed the command prefix. By default, it's the command name with as many wildcard capture groups as there are parsers, separated by the separator keyword (a space character by default).

Command Help

The help keyword specifies a help string which can be used by add_help!.

Argument Parsing

The parsers keyword specifies the parsers of the command arguments, and can contain both types and functions. If pattern contains captures, then they are run through the parsers before being passed into the handler. For repeating arguments, see Splat.

Permissions

The allowed keyword specifies Users or Roles (by ID) that are allowed to use the command. If the caller does not have permissions for the command, the fallback handler is run.

Rate Limiting

The cooldown keyword sets the rate at which a user can invoke the command. The default of nothing indicates no limit.

Fallback Function

The fallback keyword specifies a function that runs whenever a command is called but cannot be run, such as failed argument parsing, missing permissions, or rate limiting. it should accept a Client and a Message.

Additional keyword arguments are a subset of those to add_handler!.

Examples

Basic echo command with a help string:

add_command!(c, :echo, (c, m) -> reply(c, m, m.content); help="repeat a message")

The same, but excluding the command part:

add_command!(c, :echo, (c, m, msg) -> reply(c, m, msg); pattern=r"^echo (.+)")

Parsing a subtraction expression with custom parsers and separator:

add_command!(
    c, :sub, (c, m, a, b) -> reply(c, m, string(a - b));
    parsers=[Float64, Float64], separator='-',
)

Splatting some comma-separated numbers with a fallback function:

add_command!(
    c, :sum, (c, m, xs...) -> reply(c, m, string(sum(collect(xs))));
    parsers=[Splat(Float64, ',')], fallback=(c, m) -> reply(c, m, "Args must be numbers."),
)

source

add_command!(c::Client, m::Module; compile::Bool=false; kwargs...)

Add all of the bot commands defined in a module. To set up commands to be included, see @command.

source

Discord.@command — Macro.

@command name=name handler=handler kwargs...

Mark a function as a bot command to be collected by add_command! (from a module). Supported keywords are identical to add_command!.

Example

module Commands
using Discord
echo(c::Client, m::Message, noprefix::AbstractString) = reply(c, m, noprefix)
@command name=:echo handler=echo help="Echo a message" pattern=r"^echo (.+)"
end
c = Client("token")
add_command!(c, Commands)

source

Discord.delete_command! — Function.

delete_command!(c::Client, name::Symbol)

Delete a command.

source

Discord.add_help! — Function.

add_help!(
    c::Client;
    pattern::Regex=r"^help(?: (.+))?",
    help::AbstractString="Show this help message",
    nohelp::AbstractString="No help provided",
    nocmd::AbstractString="Command not found",
)

Add a help command. This can be called at any time, new commands will be included automatically.

Keywords

separator::Union{AbstractString, AbstractChar}: Separator between commands.

pattern::Regex: The command pattern (see add_command!).

help::AbstractString: Help for the help command.

nohelp::AbstractString: Help for commands without their own help string.

nocmd::AbstractString: Help for commands that aren't found.

source

Discord.set_prefix! — Function.

set_prefix!(c::Client, prefix::Union{AbstractString, AbstractChar})
set_prefix!(
    c::Client,
    prefix::Union{AbstractString, AbstractChar},
    guild::Union{Guild, Integer},
)

Set Client's command prefix. If a Guild or its ID is supplied, then the prefix only applies to that guild.

source

Discord.Splat — Type.

Splat(
    func::Base.Callable=identity,
    split::Union{AbstractString, AbstractChar}=' ',
) -> Splat

Collect a variable number of arguments from one capture group with a single parser.

source