Client

Client

Discord.Client — Type.

Client(
    token::String;
    presence::Union{Dict, NamedTuple}=Dict(),
    ttls::Dict{DataType,Union{Nothing, Period}}=Dict(),
    version::Int=6,
) -> Client

A Discord bot. Clients can connect to the gateway, respond to events, and make REST API calls to perform actions such as sending/deleting messages, kicking/banning users, etc.

Keywords

presence::Union{Dict, NamedTuple}=Dict(): Bot user's presence set upon connection. The schema here must be followed.

ttls::Dict{DataType,Union{Nothing, Period}}=Dict(): Cache lifetime overrides. Values of nothing indicate no expiry. Keys can be any of the following: Guild, DiscordChannel, Message, User, Member, or Presence. For most workloads, the defaults are sufficient.

version::Int=6: Version of the Discord API to use. Using anything but 6 is not officially supported by the Discord.jl developers.

source

Base.open — Function.

open(c::Client; delay::Period=Second(7))

Connect a Client to the Discord gateway.

The delay keyword is the time between shards connecting. It can be increased from its default if you are frequently experiencing invalid sessions upon connection.

source

Base.isopen — Function.

isopen(c::Client) -> Bool

Determine whether the Client is connected to the gateway.

source

Base.close — Function.

close(c::Client)

Disconnect the Client from the Discord gateway.

source

Base.wait — Function.

wait(c::Client)

Wait for an open Client to close.

source

Discord.me — Function.

me(c::Client) -> Union{User, Nothing}

Get the Client's bot user.

source

Event Handlers

See Events for more details.

Discord.add_handler! — Function.

add_handler!(
    c::Client,
    T::Type{<:AbstractEvent},
    func::Function;
    tag::Symbol=gensym(),
    expiry::Union{Int, Period, Nothing}=nothing,
)
add_handler!(
    func::Function;
    c::Client,
    T::Type{<:AbstractEvent},
    tag::Symbol=gensym(),
    expiry::Union{Int, Period, Nothing}=nothing,
)

Add an event handler. The handler should be a function which takes two arguments: A Client and an AbstractEvent (or a subtype). The handler is appended to the event's current handlers. You can also define a single handler for multiple event types by using a Union. do syntax is also accepted.

Keywords

tag::Symbol=gensym(): A label for the handler, which can be used to remove it with delete_handler!.

expiry::Union{Int, Period, Nothing}=nothing: The handler's expiry. If an Int is given, the handler will run that many times before expiring. If a Period is given, the handler will expire after it elapsed. The default of nothing indicates no expiry.

Note

There is no guarantee on the order in which handlers run, except that catch-all (AbstractEvent) handlers run before specific ones.

source

add_handler!(
    c::Client,
    m::Module;
    tag::Symbol=gensym(),
    expiry::Union{Int, Period, Nothing}=nothing,
)

Add all of the event handlers defined in a module. Any function you wish to use as a handler must be exported. Only functions with correct type signatures (see above) are used.

Note

If you specify a tag and/or expiry, it's applied to all of the handlers in the module. That means if you add two handlers for the same event type, one of them will be immediately overwritten.

source

Discord.delete_handler! — Function.

delete_handler!(c::Client, T::Type{<:AbstractEvent})
delete_handler!(c::Client, T::Type{<:AbstractEvent}, tag::Symbol)

Delete event handlers. If no tag is supplied, all handlers for the event are deleted. Using the tagless method is generally not recommended because it also clears default handlers which maintain the client state. If you do want to delete a default handler, use DEFAULT_HANDLER_TAG.

source

Discord.DEFAULT_HANDLER_TAG — Constant.

Tag assigned to default handlers, which you can use to delete them.

source

Bot Commands

Discord.add_command! — Function.

add_command!(
    c::Client,
    prefix::AbstractString,
    func::Function;
    tag::Symbol=gensym(),
    expiry::Union{Int, Period, Nothing}=nothing,
)
add_command!(
    func::Function;
    c::Client,
    prefix::AbstractString,
    tag::Symbol=gensym(),
    expiry::Union{Int, Period, Nothing}=nothing,
)

Add a text command handler. The handler function should take two arguments: A Client and a Message. The keyword arguments are identical to add_handler!. do syntax is also accepted.

source

Gateway Commands

Discord.request_guild_members — Function.

request_guild_members(
    c::Client,
    guilds::Union{Integer, Vector{<:Integer};
    query::AbstractString="",
    limit::Int=0,
) -> Bool

Request offline guild members of one or more Guilds. GuildMembersChunk events are sent by the gateway in response. More details here.

source

Discord.update_voice_state — Function.

update_voice_state(
    c::Client,
    guild::Integer,
    channel::Union{Integer, Nothing},
    mute::Bool,
    deaf::Bool,
) -> Bool

Join, move, or disconnect from a voice channel. A VoiceStateUpdate event is sent by the gateway in response. More details here.

source

Discord.update_status — Function.

update_status(
    c::Client,
    since::Union{Int, Nothing},
    activity::Union{Activity, Nothing},
    status::Union{PresenceStatus, AbstractString},
    afk::Bool,
) -> Bool

Indicate a presence or status update. A PresenceUpdate event is sent by the gateway in response. More details here.

source

Caching

By default, most data that comes from Discord is cached for later use. However, to avoid memory leakage, some of it is deleted after some time. The default settings are to keep everything but Messages forever, but they can be overridden in the Client constructor. Although it's not recommended, you can disable caching of certain data by clearing default handlers for relevant event types with delete_handler! and DEFAULT_HANDLER_TAG. For example, if you wanted to avoid caching any messages at all, you would delete handlers for MessageCreate and MessageUpdate events. You can also enable and disable the cache with enable_cache! and disable_cache!, which both support do syntax for temporarily altering behaviour.

Discord.enable_cache! — Function.

enable_cache!(c::Client)
enable_cache!(f::Function c::Client)

Enable the cache. do syntax is also supported.

source

Discord.disable_cache! — Function.

disable_cache!(c::Client)
disable_cache!(f::Function, c::Client)

Disable the cache. do syntax is also supported.

source

Sharding

Sharding is handled automatically: The number of available processes is the number of shards that are created. See the sharding example for more details.