Events

Events

Note that Snowflake === UInt64. Unions with Nothing indicate that a field is nullable, whereas Unions with Missing indicate that a field is optional. More details here.

Discord.AbstractEvent — Type.

An incoming event sent over the gateway. Also a catch-all event: Handlers defined on this type will execute on all events.

source

Discord.FallbackEvent — Type.

A type for defining handlers on any events which would otherwise have no non-default handler. Handlers for this type must accept an AbstractEvent.

source

Discord.UnknownEvent — Type.

An unknown event. When an event can't be parsed, due to an unknown type or any other error, it will appear as an UnknownEvent. The fields follow the schema defined here.

source

Channels

Discord.ChannelCreate — Type.

Sent when a new DiscordChannel is created.

Fields

channel :: DiscordChannel

source

Discord.ChannelUpdate — Type.

Sent when a DiscordChannel is updated.

Fields

channel :: DiscordChannel

source

Discord.ChannelDelete — Type.

Sent when a DiscordChannel is deleted.

Fields

channel :: DiscordChannel

source

Discord.ChannelPinsUpdate — Type.

Sent when a DiscordChannel's pins are updated.

Fields

channel_id         :: Snowflake
last_pin_timestamp :: Nullable{DateTime}

source

Guilds

Discord.GuildCreate — Type.

Sent when a new Guild is created.

Fields

guild :: Guild

source

Discord.GuildUpdate — Type.

Sent when a Guild is updated.

Fields

guild :: Guild

source

Discord.GuildDelete — Type.

Sent when a Guild is deleted.

Fields

guild :: AbstractGuild

source

Discord.GuildBanAdd — Type.

Sent when a User is banned from a Guild.

Fields

guild_id :: Snowflake
user     :: User

source

Discord.GuildBanRemove — Type.

Sent when a User is unbanned from a Guild.

Fields

guild_id :: Snowflake
user     :: User

source

Discord.GuildEmojisUpdate — Type.

Sent when a Guild has its Emojis updated.

Fields

guild_id :: Snowflake
emojis   :: Vector{Emoji}

source

Discord.GuildIntegrationsUpdate — Type.

Sent when a Guild has its Integrations updated.

Fields

guild_id :: Snowflake

source

Discord.GuildMemberAdd — Type.

Sent when a Member is added to a Guild.

Fields

guild_id :: Snowflake
member   :: Member

source

Discord.GuildMemberRemove — Type.

Sent when a Member is removed from a Guild.

Fields

guild_id :: Snowflake
user     :: User

source

Discord.GuildMemberUpdate — Type.

Sent when a Member is updated in a Guild.

Fields

guild_id :: Snowflake
roles    :: Vector{Snowflake}
user     :: User
nick     :: Nullable{String}

source

Discord.GuildMembersChunk — Type.

Sent when the Client requests guild members with request_guild_members.

Fields

guild_id :: Snowflake
members  :: Vector{Member}

source

Discord.GuildRoleCreate — Type.

Sent when a new Role is created in a Guild.

Fields

guild_id :: Snowflake
role     :: Role

source

Discord.GuildRoleUpdate — Type.

Sent when a Role is updated in a Guild.

Fields

guild_id :: Snowflake
role     :: Role

source

Discord.GuildRoleDelete — Type.

Sent when a Role is deleted from a Guild.

Fields

guild_id :: Snowflake
role_id  :: Snowflake

source

Messages

Discord.MessageCreate — Type.

Sent when a Message is sent.

Fields

message :: Message

source

Discord.MessageUpdate — Type.

Sent when a Message is updated.

Fields

message :: Message

source

Discord.MessageDelete — Type.

Sent when a Message is deleted.

Fields

id         :: Snowflake
channel_id :: Snowflake
guild_id   :: Optional{Snowflake}

source

Discord.MessageDeleteBulk — Type.

Sent when multiple Messages are deleted in bulk.

Fields

ids        :: Vector{Snowflake}
channel_id :: Snowflake
guild_id   :: Optional{Snowflake}

source

Discord.MessageReactionAdd — Type.

Sent when a Reaction is added to a Message.

Fields

user_id    :: Snowflake
channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}
emoji      :: Emoji

source

Discord.MessageReactionRemove — Type.

Sent when a Reaction is removed from a Message.

Fields

user_id    :: Snowflake
channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}
emoji      :: Emoji

source

Discord.MessageReactionRemoveAll — Type.

Sent when all Reactions are removed from a Message.

Fields

channel_id :: Snowflake
message_id :: Snowflake
guild_id   :: Optional{Snowflake}

source

Presence

Discord.PresenceUpdate — Type.

Sent when a User's Presence is updated.

Fields

presence :: Presence

source

Discord.TypingStart — Type.

Sent when a User begins typing.

Fields

channel_id :: Snowflake
guild_id   :: Optional{Snowflake}
user_id    :: Snowflake
timestamp  :: DateTime

source

Discord.UserUpdate — Type.

Sent when a User's details are updated.

Fields

user :: User

source

Voice

Discord.VoiceStateUpdate — Type.

Sent when a User updates their VoiceState.

Fields

state :: VoiceState

source

Discord.VoiceServerUpdate — Type.

Sent when a Guild's voice server is updated.

Fields

token    :: String
guild_id :: Snowflake
endpoint :: String

source

Webhooks

Discord.WebhooksUpdate — Type.

Sent when a DiscordChannel's Webhooks are updated.

Fields

guild_id   :: Snowflake
channel_id :: Snowflake

source

Connecting

Discord.Ready — Type.

Sent when the Client has authenticated, and contains the initial state.

Fields

v                :: Int
user             :: User
private_channels :: Vector{DiscordChannel}
guilds           :: Vector{UnavailableGuild}
session_id       :: String
_trace           :: Vector{String}

source

Discord.Resumed — Type.

Sent when a Client resumes its connection.

Fields

_trace :: Vector{String}

source