Helpers

Helpers

Discord.Permission — Type.

Bitwise permission flags. More details here.

source

Discord.has_permission — Function.

has_permission(perms::Integer, perm::Permission) -> Bool

Determine whether a bitwise OR of permissions contains one Permission.

Example

julia> has_permission(0x0420, PERM_VIEW_CHANNEL)
true

julia> has_permission(0x0420, PERM_ADMINISTRATOR)
false

source

Discord.mention — Function.

mention(x::Union{DiscordChannel, Member, Role, User}) -> String

Get the mention string for an entity.

source

Discord.reply — Function.

reply(
    c::Client,
    m::Message,
    content::Union{AbstractString, AbstractDict, NamedTuple, Embed};
    at::Bool=false,
) -> Future{Response}

Reply (send a message to the same DiscordChannel) to a Message. If at is set, then the message is prefixed with the sender's mention.

source

Discord.plaintext — Function.

plaintext(m::Message) -> String
plaintext(c::Client, m::Message) -> String

Get the Message contents with any User mentions replaced with their plaintext. If a Client is provided, DiscordChannels Role are also replaced. However, only channels and roles stored in state are replaced; no API requests are made.

source

Discord.upload_file — Function.

upload_file(c::Client, ch::DiscordChannel, path::AbstractString; kwargs...) -> Message

Send a Message with a file Attachment. Any keywords are passed on to create_message.

source

Discord.set_game — Function.

set_game(
    c::Client,
    name::AbstractString,
    type::Union{ActivityType, Int}=AT_GAME,
    since::Union{Int, Nothing}=nothing,
    status::Union{PresenceStatus, AbstractString}=PS_ONLINE,
    afk::Bool=false,
    kwargs...,
) -> Bool

Shortcut for update_status to set the Client's Activity.

source