Helpers

Helpers

Discord.reply — Function.

reply(c::Client, m::Message, content::AbstractString; at::Bool=false)

Reply (send a message to the same DiscordChannel) to a Message. If at is set, then the message is prefixed with the sender's mention.

source

Discord.mention — Function.

mention(x::Union{DiscordChannel, Member, Role, User}) -> String

Get the mention string for an entity.

source

Discord.replace_mentions — Function.

replace_mentions(m::Message) -> String

Get the Message contents with any User mentions replaced with their plaintext.

source

Discord.upload_file — Function.

upload_file(c::Client, ch::DiscordChanne, path::AbstractString; kwargs...) -> Message

Send a Message with a file Attachment. Any keywords are passed on to create_message.

source

Discord.set_game — Function.

set_game(
    c::Client,
    name::AbstractString,
    type::Union{ActivityType, Int}=AT_GAME,
    since::Union{Int, Nothing}=nothing,
    status::Union{PresenceStatus, AbstractString}=PS_ONLINE,
    afk::Bool=false,
    kwargs...,
) -> Bool

Shortcut for update_status to set the client's Activity.

source