Reference

Julicord.AbstractEvent

Julicord.Activity

Julicord.Attachment

Julicord.Ban

Julicord.Client

Julicord.DiscordChannel

Julicord.Embed

Julicord.Emoji

Julicord.Guild

Julicord.Integration

Julicord.Invite

Julicord.InviteMetadata

Julicord.Member

Julicord.Message

Julicord.Presence

Julicord.Reaction

Julicord.Role

Julicord.UnavailableGuild

Julicord.UnknownEvent

Julicord.User

Julicord.VoiceState

Julicord.Webhook

Base.open

Base.wait

Julicord.add_handler!

Julicord.clear_handlers!

Julicord.me

Julicord.request_guild_members

Julicord.state

Julicord.update_status

Julicord.update_voice_state

Julicord.AbstractEvent — Type.

An incoming event sent over the gateway. Also a catch-all event: Handlers defined on this type will execute on all events.

source

Julicord.Activity — Method.

A user activity. More details here.

source

Julicord.Client — Type.

Client(token::String) -> Client

A Discord bot.

source

Julicord.UnknownEvent — Type.

An unknown event.

source

Julicord.add_handler! — Method.

add_handler!(c::Client, evt::Type{<:AbstractEvent}, func::Function)

Add a handler for the given event type. The handler should be a function which takes two arguments: A Client and an AbstractEvent (or a subtype). The handler is appended the event's current handlers.

Note

The set of handlers for a given event is stored as a Set{Function}. This protects against adding duplicate handlers, except when you pass an anonymous function. Therefore, it's recommended to define your handler functions beforehand.

Also note that there is no guarantee on the order in which handlers run.

source

Julicord.clear_handlers! — Method.

clear_handlers!(c::Client, evt::Type{<:AbstractEvent})

Removes all handlers for the given event type.

source

Julicord.me — Method.

me(c::Client) -> Union{User, Nothing}

Get the client's bot user.

source

Julicord.request_guild_members — Method.

request_guild_members(
    c::Client,
    guild_id::Union{Snowflake, Vector{Snowflake};
    query::AbstractString="",
    limit::Int=0,
) -> Bool

Request offline guild members of one or more guilds. More details here.

source

Julicord.state — Method.

state(c::Client) -> State

Get the client state.

source

Julicord.update_status — Method.

update_status(
    c::Client,
    since::Union{Int, Nothing},
    activity::Union{Activity, Nothing},
    status::PresenceStatus,
    afk::Bool,
) -> Bool

Indicate a presence or status update. More details here.

source

Julicord.Attachment — Method.

A message attachment. More details here.

source

Julicord.Ban — Method.

A user ban. More details here.

source

Julicord.DiscordChannel — Method.

A Discord channel. More details here. Note: The name Channel is already used, hence the prefix.

source

Julicord.Embed — Method.

A message embed. More details here.

source

Julicord.Emoji — Method.

An emoji. More details here.

source

Julicord.Guild — Method.

A guild (server). More details here.

source

Julicord.Integration — Method.

A server integration. More details here.

source

Julicord.Invite — Method.

An invite to a guild. More details here.

source

Julicord.InviteMetadata — Method.

Metadata for an Invite. More details here.

source

Julicord.Member — Method.

A guild member. More details here.

source

Julicord.Message — Method.

A message. More details here.

source

Julicord.Presence — Method.

A user presence. More details here.

source

Julicord.Reaction — Method.

A reaction. More details here.

source

Julicord.Role — Method.

A user role. More details here.

source

Julicord.UnavailableGuild — Method.

An unavailable guild (server). More details here.

source

Julicord.User — Method.

A Discord user. More details here.

source

Julicord.VoiceState — Method.

A users's voice connection status. More details here.

source

Julicord.Webhook — Method.

A Webhook. More details here.

source

Base.open — Method.

open(c::Client)

Log in to the Discord gateway and begin responding to events.

source

Base.wait — Method.

wait(c::Client)

Wait for an open client to close.

source

Julicord.update_voice_state — Method.

update_voice_state(
    c::Client,
    guild_id::Snowflake,
    channel_id::Union{Snowflake, Nothing},
    self_mute::Bool,
    self_deaf::Bool,
) -> Bool

Join, move, or disconnect from a voice channel. More details here.

source