Reference

Julicord.AbstractEvent

Julicord.AbstractGuild

Julicord.Activity

Julicord.Attachment

Julicord.Ban

Julicord.Client

Julicord.DiscordChannel

Julicord.Embed

Julicord.Emoji

Julicord.Guild

Julicord.Integration

Julicord.Invite

Julicord.InviteMetadata

Julicord.Member

Julicord.Message

Julicord.OnLimit

Julicord.Overwrite

Julicord.Presence

Julicord.Reaction

Julicord.Response

Julicord.Role

Julicord.UnavailableGuild

Julicord.UnknownEvent

Julicord.User

Julicord.VoiceState

Julicord.Webhook

Base.open

Base.wait

Julicord.add_handler!

Julicord.add_role

Julicord.ban_member

Julicord.bulk_delete

Julicord.clear_handlers!

Julicord.create_invite

Julicord.create_webhook

Julicord.delete

Julicord.delete_channel

Julicord.delete_handler!

Julicord.delete_integration

Julicord.delete_invite

Julicord.delete_overwrite

Julicord.delete_reactions

Julicord.delete_role

Julicord.delete_webhook

Julicord.edit

Julicord.execute_github

Julicord.execute_slack

Julicord.execute_webhook

Julicord.get_invite

Julicord.get_invites

Julicord.get_member

Julicord.get_message

Julicord.get_messages

Julicord.get_pinned_messages

Julicord.get_reactions

Julicord.get_webhook

Julicord.get_webhooks

Julicord.kick_member

Julicord.me

Julicord.modify_channel

Julicord.modify_integration

Julicord.modify_member

Julicord.modify_overwrite

Julicord.modify_role

Julicord.modify_webhook

Julicord.pin

Julicord.react

Julicord.remove_role

Julicord.reply

Julicord.request_guild_members

Julicord.send_message

Julicord.state

Julicord.sync_integration

Julicord.trigger_typing

Julicord.unpin

Julicord.update_status

Julicord.update_voice_state

Julicord.AbstractEvent — Type.

An incoming event sent over the gateway. Also a catch-all event: Handlers defined on this type will execute on all events.

source

Julicord.Activity — Method.

A user activity. More details here.

source

Julicord.Client — Type.

Client(token::String; on_limit::OnLimit=LIMIT_IGNORE, ttl::Period=Hour(1) -> Client

A Discord bot.

Keywords

on_limit::OnLimit=LIMIT_IGNORE: Client's behaviour when it hits a rate limit.

ttl::Period=Hour(1) Amount of time that cache entries are kept.

version::Int=6: Version of Discord API to use. Using anything but 6 is not supported.

source

Julicord.UnknownEvent — Type.

An unknown event.

source

Julicord.add_handler! — Method.

add_handler!(
    c::Client,
    evt::Type{<:AbstractEvent},
    func::Function;
    tag::Symbol=gensym(),
    expiry::Union{Int, Period}=-1,
)

Add a handler for the given event type. The handler should be a function which takes two arguments: A Client and an AbstractEvent (or a subtype). The handler is appended the event's current handlers.

Keywords

tag::Symbol=gensym(): A label for the handler, which can be used to remove it with delete_handler!.

expiry::Union{Int, Period}=-1: The handler's expiry. If an Int is given, the handler

will run a set number of times before expiring. If a Period is given, the handler will expire after that amount of time has elapsed. The default of -1 indicates no expiry.

Note

There is no guarantee on the order in which handlers run, except that catch-all handlers run before specific ones.

source

Julicord.add_role — Method.

add_role(
    c::Client,
    guild::Union{AbstractGuild, Integer},
    user::Union{User, Integer},
    role::Union{Role, Integer},
) -> Response

Add a Role to a Member in an AbstractGuild..

source

Julicord.ban_member — Method.

ban_member(
    c::Client,
    guild::Union{Guild, Integer},
    user::Union{User, Integer};
    params...,
) -> Response

Ban a Member from an AbstractGuild.

Keywords

delete_message_days::Integer: Number of days to delete the messages for (0-7).

reason::AbstractString: Reason for the ban.

source

Julicord.bulk_delete — Method.

bulk_delete(
    c::Client,
    channel::Union{DiscordChannel, Integer},
    messages::Union{Vector{Message}, Vector{<:Integer}},
) -> Response

Delete multiple Messages from a DiscordChannel.

source

Julicord.clear_handlers! — Method.

clear_handlers!(c::Client, evt::Type{<:AbstractEvent})

Remove all handlers for the given event type. Using this is generally not recommended because it also clears default handlers which maintain the client state. Instead, try to add handlers with specific tags and delete them with delete_handler!.

source

Julicord.create_invite — Method.

create_invite(
    c::Client,
    channel::Union{DiscordChannel, Integer};
    params...,
) -> Response{Invite}

Create an Invite to a DiscordChannel.

Keywords

max_uses::Int: Max number of uses (0 if unlimited).

max_age::Int: Duration in seconds before expiry (0 if never).

temporary::Bool: Whether this invite only grants temporary membership.

unique::Bool: Whether not to try to reuse a similar invite.

More details here.

source

Julicord.create_webhook — Method.

create_webhook(
    c::Client,
    channel::Union{DiscordChannel, Integer},
    params...,
) -> Response

Create a Webhook in a DiscordChannel.

Keywords

name::AbstractString - name of the webhook (2-23 characters)

avatar::AbstractString - image for the default webhook avatar

More details here.

source

Julicord.delete — Method.

delete(c::Client, m::Message) -> Response

Delete a Message.

source

Julicord.delete_channel — Method.

delete_channel(c::Client, channel:::Union{DiscordChannel, Integer}) -> Response{Channel}

Delete a DiscordChannel.

source

Julicord.delete_handler! — Method.

delete_handler!(c::Client, evt::Type{<:AbstractEvent}, tag::Symbol)

Delete a single handler by event type and tag.

source

Julicord.delete_integration — Method.

delete_integration(
    c::Client,
    guild::Union{AbstractGuild, Integer},
    integration::Union{Integration, Integer},
) -> Response

Delete an Integration in an AbstractGuild.

source

Julicord.delete_invite — Method.

delete_invite(c::Client, invite::Union{Invite, AbstractString}) -> Response{Invite}

Delete an Invite.

source

Julicord.delete_overwrite — Method.

delete_overwrite(
    c::Client,
    overwrite::Union{Overwrite, Integer},
    channel::Union{DiscordChannel, Integer}
) -> Response{Overwrite}

Delete an Overwrite in a DiscordChannel.

source

Julicord.delete_reactions — Method.

delete_reactions(c::Client, m::Message) -> Response

Delete all the reactions from a Message.

source

Julicord.delete_role — Method.

delete_role(
    c::Client,
    role::Union{Role, Integer},
    guild::Union{AbstractGuild, Integer},
) -> Response

Modify a Role in a DiscordChannel.

source

Julicord.delete_webhook — Method.

delete_webhook(c::Client, webhook::Union{Webhook, Integer}) -> Response
delete_webhook(
    c::Client,
    webhook::Union{Webhook, Integer},
    token::AbstractString,
) -> Response

Delete a Webhook.

source

Julicord.edit — Method.

edit(c::Client, m::Message, content::Union{AbstractString, Dict}) -> Response{Message}

Edit a Message.

source

Julicord.execute_github — Method.

execute_github(
    c::Client,
    webhook::Union{Webhook, Integer},
    token::AbstractString;
    wait::Bool=true,
    params...,
) -> Union{Response{Message}, Response{Nothing}}

Execute a Github Webhook.

More details here.

source

Julicord.execute_slack — Method.

execute_slack(
    c::Client,
    webhook::Union{Webhook, Integer},
    token::AbstractString;
    wait::Bool=true,
    params...,
) -> Union{Response{Message}, Response}

Execute a Slack Webhook.

More details here.

source

Julicord.execute_webhook — Method.

execute_webhook(
    c::Client,
    webhook::Union{Webhook, Integer},
    token::AbstractString;
    wait::Bool=false,
    params...,
) -> Union{Response{Message}, Response{Nothing}}

Execute a Webhook. If wait is set, the created Message is returned.

Keywords

content::AbstractString: The message contents (up to 2000 characters).

username::AbstractString: Override the default username of the webhook.

avatar_url::AbstractString: Override the default avatar of the webhook.

tts::Bool: Whether this is a TTS message.

file::AbstractDict: The contents of the file being sent.

embeds::AbstractDict: Embedded rich content.

More details here.

source

Julicord.get_invite — Method.

get_invite(
    c::Client,
    invite::Union{Invite, AbstractString};
    with_counts::Bool=false,
) -> Response{Invite}

Get an Invite. If with_counts is set, the Invite will contain approximate member counts.

source

Julicord.get_invites — Method.

get_invites(
    c::Client,
    channel::Union{DiscordChannel, Integer},
) -> Response{Vector{Invite}}

Get a list of Invites from a DiscordChannel.

source

Julicord.get_member — Method.

get_member(
    c::Client,
    guild::Union{AbstractGuild, Integer},
    user::Union{User, Integer},
) -> Response{Member}

Get a Member in an AbstractGuild.

source

Julicord.get_message — Method.

get_message(
    c::Client,
    channel::Union{DiscordChannel, Integer},
    message::Integer,
) -> Response{Message}
get_message(c::Client, m::Message) -> Response{Message}

Get a Message from a DiscordChannel.

source

Julicord.get_messages — Method.

get_messages(
    c::Client,
    channel::Union{DiscordChannel, Integer};
    params...,
) -> Response{Vector{Message}}

Get a list of Messages from a DiscordChannel.

Keywords

around::Integer: Get messages around this message ID.

before::Integer: Get messages before this message ID.

after::Integer: Get messages after this message ID.

limit::Int: Maximum number of messages.

More details here.

source

Julicord.get_pinned_messages — Method.

get_pinned_messages(
    c::Client,
    channel::Union{DiscordChannel, Integer},
) -> Response{Vector{Message}}

Get a list of Messages pinned in a DiscordChannel.

source

Julicord.get_reactions — Method.

get_reactions(
    c::Client,
    m::Message,
    emoji::Union{Emoji, AbstractString},
) -> Response{Vector{User}}

Get the users who reacted to a Message with an Emoji.

source

Julicord.get_webhook — Method.

get_webhook(c::Client, webhook::Union{Webhook, Integer}) -> Response{Webhook}
get_webhook(
    c::Client,
    webhook::Union{Webhook, Integer},
    token::AbstractString,
) -> Response{Webhook}

Get a Webhook.

source

Julicord.get_webhooks — Method.

get_webhooks(
    c::Client,
    channel::Union{DiscordChannel, Integer},
) -> Response{Vector{Webhook}}

Get a list of Webhooks from a DiscordChannel.

source

Julicord.kick_member — Method.

kick_member(
    c::Client,
    guild::Union{AbstractGuild, Integer},
    user::Union{User, Integer},
) -> Response

Kick a Member from an AbstractGuild.

source

Julicord.me — Method.

me(c::Client) -> Union{User, Nothing}

Get the client's bot user.

source

Julicord.modify_channel — Method.

modify_channel(
    c::Client,
    channel::Union{DiscordChannel, Integer};
    params...,
) -> Response{Channel}

Modify a DiscordChannel.

Keywords

name::AbstractString: Channel name (2-100 characters).

topic::AbstractString: Channel topic (up to 1024 characters).

nsfw::Bool: Whether the channel is NSFW.

rate_limit_per_user::Int: Seconds a user must wait before sending another message.

position::Int The position in the left-hand listing.

bitrate::Int The bitrate in bits of the voice channel.

user_limit::Int: The user limit of the voice channel.

permission_overwrites::Vector{<:AbstractDict}: Channel or category-specific permissions.

parent_id::Integer: ID of the new parent category.

More details here.

source

Julicord.modify_integration — Method.

modify_integration(
    c::Client,
    guild::Union{AbstractGuild, Integer},
    integration::Union{Integration, Integer};
    params...,
) -> Response{Integration}

Modify an Integration in an AbstractGuild.

Keywords

expire_behavior::Integer: The behavior when an integration subscription lapses.

expire_grace_period::Integer: Period (in seconds) where the integration will ignore lapsed subscriptions.

enable_emoticons::Bool: Whether emoticons should be synced for this integration (Twitch only currently).

source

Julicord.modify_member — Method.

modify_member(
    c::Client,
    guild::Union{AbstractGuild, Integer},
    user::Union{User, Integer};
    params...,
) -> Response{Member}

Modify a Member in an AbstractGuild.

Keywords

nick::AbstractString: Value to set the member's nickname to.

roles::Vector: List of role ids the member is assigned.

mute::Bool: Whether the member should be muted.

deaf::Bool: Whether the member should be deafened.

channel_id::Integer: ID of a voice channel to move the member to.

More details here.

source

Julicord.modify_overwrite — Method.

modify_overwrite(
    c::Client,
    overwrite::Union{Overwrite, Integer},
    channel::Union{DiscordChannel, Integer};
    params...,
) -> Response{Overwrite}

Modify an Overwrite in a DiscordChannel.

Keywords

allow::Int: the bitwise OR of the allowed permissions.

deny::Int: the bitwise OR of the denied permissions.

type::AbstractString: "member" for a user or "role" for a role.

More details here.

source

Julicord.modify_role — Method.

modify_role(
    c::Client,
    role::Union{Role, Integer},
    guild::Union{AbstractGuild, Integer};
    params...,
) -> Response{Role}

Modify a Role in a DiscordChannel.

Keywords

name::AbstractString: Name of the role.

permissions::Int: Bitwise OR of the enabled/disabled permissions.

color::Int: RGB color value.

hoist::Bool: Whether the role should be displayed separately in the sidebar.

mentionable::Bool: Whether the role should be mentionable.

More details here.

source

Julicord.modify_webhook — Method.

modify_webhook(c::Client, webhook::Union{Webhook, Integer}; params...) -> Response{Webhook}
modify_webhook(
    c::Client,
    webhook::Union{Webhook, Integer},
    token::AbstractString;
    params...,
) -> Response{Webhook}

Modify a Webhook.

Keywords

name::AbstractString: Name of the webhook.

avatar::AbstractString: Avatar data string.

channel_id::Integer: The channel this webhook should be moved to.

If using a token, channel_id cannot be used and the returned Webhook will not contain a User.

More details here.

source

Julicord.pin — Method.

pin(c::Client, m::Message) -> Response

Pin a Message.

source

Julicord.react — Method.

react(c::Client, m::Message, emoji::AbstractString) -> Response

React to a Message.

source

Julicord.remove_role — Method.

remove_role(
    c::Client,
    guild::Union{AbstractGuild, Integer},
    user::Union{User, Integer},
    role::Union{Role, Integer},
) -> Response

Remove a Role from a Member in an AbstractGuild.

source

Julicord.reply — Method.

reply(c::Client, m::Message, content::Union{AbstractString, Dict}) -> Response{Message}

Reply to a Message (send a message to the same channel).

source

Julicord.request_guild_members — Method.

request_guild_members(
    c::Client,
    guild_id::Union{Snowflake, Vector{Snowflake};
    query::AbstractString="",
    limit::Int=0,
) -> Bool

Request offline guild members of one or more guilds. More details here.

source

Julicord.send_message — Method.

send_message(
    c::Client,
    channel::Union{DiscordChannel, Integer},
    content::Union{AbstractString, AbstractDict},
) -> Response{Message}

Send a Message to a DiscordChannel.

source

Julicord.state — Method.

state(c::Client) -> State

Get the client state.

source

Julicord.sync_integration — Method.

sync_integration(
    c::Client,
    guild::Union{Guild, Integer},
    integration::Union{Integration, Integer},
) -> Response

Sync an Integration in an AbstractGuild.

source

Julicord.trigger_typing — Method.

trigger_typing(c::Client, channel::Union{DiscordChannel, Integer}) -> Response

Trigger the typing indicator in a DiscordChannel.

source

Julicord.unpin — Method.

unpin(c::Client, m::Message) -> Response

Unpin a Message.

source

Julicord.update_status — Method.

update_status(
    c::Client,
    since::Union{Int, Nothing},
    activity::Union{Activity, Nothing},
    status::PresenceStatus,
    afk::Bool,
) -> Bool

Indicate a presence or status update. More details here.

source

Julicord.AbstractGuild — Type.

A guild (server). Can either be an UnavailableGuild or a Guild.

source

Julicord.Attachment — Method.

A message attachment. More details here.

source

Julicord.Ban — Method.

A user ban. More details here.

source

Julicord.DiscordChannel — Method.

A Discord channel. More details here. Note: The name Channel is already used, hence the prefix.

source

Julicord.Embed — Method.

A message embed. More details here.

source

Julicord.Emoji — Method.

An emoji. More details here.

source

Julicord.Guild — Method.

A guild (server). More details here.

source

Julicord.Integration — Method.

A server integration. More details here.

source

Julicord.Invite — Method.

An invite to a guild. More details here.

source

Julicord.InviteMetadata — Method.

Metadata for an Invite. More details here.

source

Julicord.Member — Method.

A guild member. More details here.

source

Julicord.Message — Method.

A message. More details here.

source

Julicord.OnLimit — Type.

Determines the behaviour of a Client when it hits a rate limit. If set to LIMIT_IGNORE, a Response is returned immediately with rate_limited set to true. If set to LIMIT_WAIT, the client blocks until the rate limit resets, then retries the request.

source

Julicord.Overwrite — Method.

An Overwrite. More details here.

source

Julicord.Presence — Method.

A user presence. More details here.

source

Julicord.Reaction — Method.

A reaction. More details here.

source

Julicord.Response — Type.

A wrapper around a response from the REST API.

Fields

val::Union{T, Nothing}: The object contained in the HTTP response. For example, a call to get_message will return a Response{Message} for which this value is a Message. If success is false, it is nothing.

success::Bool: The success state of the request. If this is true, then it is safe to access val.

cache_hit::Bool: Whether val came from the cache.

rate_limited::Bool: Whether the request was rate limited.

http_response::Union{HTTP.Messages.Response, Nothing}: The underlying HTTP response. If success is true, it is nothing.

source

Julicord.Role — Method.

A user role. More details here.

source

Julicord.UnavailableGuild — Method.

An unavailable guild (server). More details here.

source

Julicord.User — Method.

A Discord user. More details here.

source

Julicord.VoiceState — Method.

A users's voice connection status. More details here.

source

Julicord.Webhook — Method.

A Webhook. More details here.

source

Base.open — Method.

open(c::Client)

Log in to the Discord gateway and begin responding to events.

source

Base.wait — Method.

wait(c::Client)

Wait for an open client to close.

source

Julicord.update_voice_state — Method.

update_voice_state(
    c::Client,
    guild_id::Snowflake,
    channel_id::Union{Snowflake, Nothing},
    self_mute::Bool,
    self_deaf::Bool,
) -> Bool

Join, move, or disconnect from a voice channel. More details here.

source