REST API

Response

Discord.Response — Type.

A wrapper around a response from the REST API. Every function which wraps a Discord REST API endpoint returns a value of this type.

Fields

val::Union{T, Nothing}: The object contained in the HTTP response. For example, a call to get_message will return a Response{Message} for which this value is a Message. If success is false, it is nothing.

success::Bool: The success state of the request. If this is true, then it is safe to access val.

cache_hit::Bool: Whether val came from the cache.

rate_limited::Bool: Whether the request was rate limited.

http_response::Union{HTTP.Messages.Response, Nothing}: The underlying HTTP response. If no HTTP request was made (cache hit, rate limit, etc.), it is nothing.

source

Rate Limiting

Discord.OnLimit — Type.

Passed as a keyword argument to Client to determine the client's behaviour when it hits a rate limit. If set to LIMIT_IGNORE, a Response is returned immediately with rate_limited set to true. If set to LIMIT_WAIT, the client blocks until the rate limit resets, then retries the request.

source

Endpoints

Discord.get_message — Function.

get_message(
    c::Client,
    channel::Union{DiscordChannel, Integer},
    message::Integer,
) -> Response{Message}
get_message(c::Client, m::Message) -> Response{Message}

Get a Message from a DiscordChannel.

source

TODO