Types

Types

This page is organized in mostly-alphabetical order.

Discord.Activity — Type.

A User activity. More details here.

source

Discord.ActivityTimestamps — Type.

Indicates the start and stop of an Activity. More details here.

source

Discord.ActivityParty — Type.

The current party of an Activity's player. More details here.

source

Discord.ActivityAssets — Type.

Images and hover text for an Activity. More details here.

source

Discord.ActivitySecrets — Type.

Secrets for Rich Presence joining and spectating of an Activity. More details here.

source

Discord.ActivityType — Type.

The type of an Activity. More details here.

source

Discord.ActivityFlags — Type.

Flags which indicate what an Activity payload contains. More details here.

source

Discord.Attachment — Type.

A Message attachment. More details here.

source

Discord.AuditLog — Type.

An audit log. More details here.

source

Discord.AuditLogEntry — Type.

An entry in an AuditLog. More details here.

source

Discord.AuditLogChange — Type.

A change item in an AuditLogEntry.

The first type parameter is the type of new_value and old_value. The second is the type of the entity that new_value and old_value belong(ed) to.

More details here.

source

Discord.AuditLogOptions — Type.

Optional information in an AuditLogEntry.

source

Discord.ActionType — Type.

AuditLog action types. More details here.

source

Discord.Ban — Type.

A User ban. More details here.

source

Discord.DiscordChannel — Type.

A Discord channel. More details here. Note: The name Channel is already used, hence the prefix.

source

Discord.ChannelType — Type.

The type of a DiscordChannel.

source

Discord.Connection — Type.

<<<<<<< HEAD A User Connection. ======= A User connection to an external service (Twitch, YouTube, etc.).

master

More details here.

source

Discord.Embed — Type.

A Message embed. More details here.

source

Discord.EmbedThumbnail — Type.

An Embed's thumbnail image information. More details here.

source

Discord.EmbedVideo — Type.

An Embed's video information. More details here.

source

Discord.EmbedImage — Type.

An Embed's image information. More details here.

source

Discord.EmbedProvider — Type.

An Embed's provider information. More details here.

source

Discord.EmbedAuthor — Type.

An Embed's author information. More details here.

source

Discord.EmbedFooter — Type.

An Embed's footer information. More details here.

source

Discord.EmbedField — Type.

An Embed field. More details here.

source

Discord.Emoji — Type.

An emoji. More details here.

source

Discord.AbstractGuild — Type.

A guild (server). Can either be an UnavailableGuild or a Guild.

source

Discord.Guild — Type.

A guild (server). More details here.

source

Discord.UnavailableGuild — Type.

An unavailable guild (server). More details here.

source

Discord.VerificationLevel — Type.

A Guild's verification level. More details here.

source

Discord.MessageNotificationLevel — Type.

A Guild's default message notification level. More details here.

source

Discord.ExplicitContentFilterLevel — Type.

A Guild's explicit content filter level. More details here.

source

Discord.MFALevel — Type.

A Guild's MFA level. More details here.

source

Discord.GuildEmbed — Type.

A Guild embed. More details here.

source

Discord.Integration — Type.

A Guild integration. More details here.

source

Discord.IntegrationAccount — Type.

An Integration account. More details here.

source

Discord.Invite — Type.

An invite to a Guild. More details here.

source

Discord.InviteMetadata — Type.

Metadata for an Invite. More details here.

source

Discord.Member — Type.

A Guild member. More details here.

source

Discord.Message — Type.

A message. More details here.

source

Discord.MessageActivity — Type.

A Message activity. More details here.

source

Discord.MessageApplication — Type.

A Rich Presence Message's application information. More details here.

source

Discord.MessageType — Type.

The type of a Message. More details here.

source

Discord.MessageActivityType — Type.

The type of a Message activity. More details here.

source

Discord.Overwrite — Type.

A permission overwrite. More details here.

source

Discord.Presence — Type.

A User's presence. More details here.

source

Discord.PresenceStatus — Type.

A User's status sent in a Presence. More details here.

source

Discord.Reaction — Type.

A Message reaction. More details here.

source

Discord.Role — Type.

A User role. More details here.

source

Discord.User — Type.

A Discord user. More details here.

source

Discord.VoiceRegion — Type.

A region for a Guild's voice server. More details here.

source

Discord.VoiceState — Type.

A User's voice connection status. More details here.

source

Discord.Webhook — Type.

A Webhook. More details here.

source