Types

Types

This page is organized in mostly-alphabetical order. Note that Snowflake === UInt64. Unions with Nothing indicate that a field is nullable, whereas Unions with Missing indicate that a field is optional. More details here.

Most of the time, you'll receive objects of these types as return values rather than creating them yourself. However, should you wish to create your own instances from scratch, all of these types have keyword constructors. If a field value can be missing, then its keyword is optional.

Discord.Activity — Type.

A User activity. More details here.

Fields

name           :: String
type           :: ActivityType
url            :: Union{String, Missing, Nothing}
timestamps     :: Union{ActivityTimestamps, Missing}
application_id :: Union{Snowflake, Missing}
details        :: Union{String, Missing, Nothing}
state          :: Union{String, Missing, Nothing}
party          :: Union{ActivityParty, Missing}
assets         :: Union{ActivityAssets, Missing}
secrets        :: Union{ActivitySecrets, Missing}
instance       :: Union{Bool, Missing}
flags          :: Union{Int, Missing}

source

Discord.ActivityTimestamps — Type.

The start and stop times of an Activity. More details here.

Fields

start :: Union{DateTime, Missing}
stop  :: Union{DateTime, Missing}

source

Discord.ActivityParty — Type.

The current party of an Activity's player. More details here.

Fields

id   :: Union{String, Missing}
size :: Union{Vector{Int}, Missing}

source

Discord.ActivityAssets — Type.

Images and hover text for an Activity. More details here.

Fields

large_image :: Union{String, Missing}
large_text  :: Union{String, Missing}
small_image :: Union{String, Missing}
small_text  :: Union{String, Missing}

source

Discord.ActivitySecrets — Type.

Secrets for Rich Presence joining and spectating of an Activity. More details here.

Fields

join     :: Union{String, Missing}
spectate :: Union{String, Missing}
match    :: Union{String, Missing}

source

Discord.ActivityType — Type.

An Activity's type. Available values are AT_GAME, AT_STREAMING, AT_LISTENING, and AT_WATCHING. More details here.

source

Discord.ActivityFlags — Type.

Flags which indicate what an Activity payload contains. More details here.

source

Discord.Attachment — Type.

A Message attachment. More details here.

Fields

id        :: Snowflake
filename  :: String
size      :: Int
url       :: String
proxy_url :: String
height    :: Union{Int, Missing}
width     :: Union{Int, Missing}

source

Discord.AuditLog — Type.

An audit log. More details here.

Fields

webhooks          :: Vector{Webhook}
users             :: Vector{User}
audit_log_entries :: Vector{AuditLogEntry}

source

Discord.AuditLogEntry — Type.

An entry in an AuditLog. More details here.

Fields

target_id   :: Union{Snowflake, Nothing}
changes     :: Union{Vector{AuditLogChange}, Missing}
user_id     :: Snowflake
id          :: Snowflake
action_type :: ActionType
options     :: Union{AuditLogOptions, Missing}
reason      :: Union{String, Missing}

source

Discord.AuditLogChange — Type.

A change item in an AuditLogEntry.

The first type parameter is the type of new_value and old_value. The second is the type of the entity that new_value and old_value belong(ed) to.

More details here.

Fields

new_value :: Union{T, Missing} where T
old_value :: Union{T, Missing} where T
key       :: String
type      :: Type{U} where U

source

Discord.AuditLogOptions — Type.

Optional information in an AuditLogEntry. More details here.

Fields

delete_member_days :: Union{Int, Missing}
members_removed    :: Union{Int, Missing}
channel_id         :: Union{Snowflake, Missing}
count              :: Union{Int, Missing}
id                 :: Union{Snowflake, Missing}
type               :: Union{OverwriteType, Missing}
role_name          :: Union{String, Missing}

source

Discord.ActionType — Type.

AuditLog action types. More details here.

source

Discord.Ban — Type.

A User ban. More details here.

Fields

reason :: Union{String, Nothing}
user   :: User

source

Discord.DiscordChannel — Type.

A Discord channel. More details here.

Note: The name Channel is already used, hence the prefix.

Fields

id                    :: Snowflake
type                  :: ChannelType
guild_id              :: Union{Snowflake, Missing}
position              :: Union{Int, Missing}
permission_overwrites :: Union{Vector{Overwrite}, Missing}
name                  :: Union{String, Missing}
topic                 :: Union{String, Missing, Nothing}
nsfw                  :: Union{Bool, Missing}
last_message_id       :: Union{Snowflake, Missing, Nothing}
bitrate               :: Union{Int, Missing}
user_limit            :: Union{Int, Missing}
rate_limit_per_user   :: Union{Int, Missing}
recipients            :: Union{Vector{User}, Missing}
icon                  :: Union{String, Missing, Nothing}
owner_id              :: Union{Snowflake, Missing}
application_id        :: Union{Snowflake, Missing}
parent_id             :: Union{Snowflake, Missing, Nothing}
last_pin_timestamp    :: Union{DateTime, Missing, Nothing}

source

Discord.ChannelType — Type.

A DiscordChannel's type. Available values are CT_GUILD_TEXT, CT_DM, CT_GUILD_VOICE, CT_GROUP_DM, and CT_GUILD_CATEGORY.

source

Discord.Connection — Type.

A User connection to an external service (Twitch, YouTube, etc.). More details here.

Fields

id           :: String
name         :: String
type         :: String
revoked      :: Bool
integrations :: Vector{Integration}

source

Discord.Embed — Type.

A Message embed. More details here.

Fields

title       :: Union{String, Missing}
type        :: Union{String, Missing}
description :: Union{String, Missing}
url         :: Union{String, Missing}
timestamp   :: Union{DateTime, Missing}
color       :: Union{Int, Missing}
footer      :: Union{EmbedFooter, Missing}
image       :: Union{EmbedImage, Missing}
thumbnail   :: Union{EmbedThumbnail, Missing}
video       :: Union{EmbedVideo, Missing}
provider    :: Union{EmbedProvider, Missing}
author      :: Union{EmbedAuthor, Missing}
fields      :: Union{Vector{EmbedField}, Missing}

source

Discord.EmbedThumbnail — Type.

An Embed's thumbnail image information. More details here.

Fields

url       :: Union{String, Missing}
proxy_url :: Union{String, Missing}
height    :: Union{Int, Missing}
width     :: Union{Int, Missing}

source

Discord.EmbedVideo — Type.

An Embed's video information. More details here.

Fields

url    :: Union{String, Missing}
height :: Union{Int, Missing}
width  :: Union{Int, Missing}

source

Discord.EmbedImage — Type.

An Embed's image information. More details here.

Fields

url       :: Union{String, Missing}
proxy_url :: Union{String, Missing}
height    :: Union{Int, Missing}
width     :: Union{Int, Missing}

source

Discord.EmbedProvider — Type.

An Embed's provider information. More details here.

Fields

name :: Union{String, Missing}
url  :: Union{String, Missing, Nothing}

source

Discord.EmbedAuthor — Type.

An Embed's author information. More details here.

Fields

name           :: Union{String, Missing}
url            :: Union{String, Missing}
icon_url       :: Union{String, Missing}
proxy_icon_url :: Union{String, Missing}

source

Discord.EmbedFooter — Type.

An Embed's footer information. More details here.

Fields

text           :: String
icon_url       :: Union{String, Missing}
proxy_icon_url :: Union{String, Missing}

source

Discord.EmbedField — Type.

An Embed field. More details here.

Fields

name   :: String
value  :: String
inline :: Union{Bool, Missing}

source

Discord.Emoji — Type.

An emoji. More details here.

Fields

id             :: Union{Snowflake, Nothing}
name           :: String
roles          :: Union{Vector{Snowflake}, Missing}
user           :: Union{User, Missing}
require_colons :: Union{Bool, Missing}
managed        :: Union{Bool, Missing}
animated       :: Union{Bool, Missing}

source

Discord.AbstractGuild — Type.

A Discord guild (server). Can either be an UnavailableGuild or a Guild.

source

Discord.Guild — Type.

A Discord guild (server). More details here.

Fields

id                            :: Snowflake
name                          :: String
icon                          :: Union{String, Nothing}
splash                        :: Union{String, Nothing}
owner                         :: Union{Bool, Missing}
owner_id                      :: Union{Snowflake, Missing}
permissions                   :: Union{Int, Missing}
region                        :: Union{String, Missing}
afk_channel_id                :: Union{Snowflake, Missing, Nothing}
afk_timeout                   :: Union{Int, Missing}
embed_enabled                 :: Union{Bool, Missing}
embed_channel_id              :: Union{Snowflake, Missing, Nothing}
verification_level            :: VerificationLevel
default_message_notifications :: Union{MessageNotificationLevel, Missing}
explicit_content_filter       :: Union{ExplicitContentFilterLevel, Missing}
roles                         :: Union{Vector{Role}, Missing}
emojis                        :: Union{Vector{Emoji}, Missing}
features                      :: Vector{String}
mfa_level                     :: Union{MFALevel, Missing}
application_id                :: Union{Snowflake, Missing, Nothing}
widget_enabled                :: Union{Bool, Missing}
widget_channel_id             :: Union{Snowflake, Missing, Nothing}
system_channel_id             :: Union{Snowflake, Missing, Nothing}
joined_at                     :: Union{DateTime, Missing}
large                         :: Union{Bool, Missing}
unavailable                   :: Union{Bool, Missing}
member_count                  :: Union{Int, Missing}
voice_states                  :: Union{Vector{VoiceState}, Missing}
members                       :: Union{Vector{Member}, Missing}
channels                      :: Union{Vector{DiscordChannel}, Missing}
presences                     :: Union{Vector{Presence}, Missing}

source

Discord.UnavailableGuild — Type.

An unavailable Discord guild (server). More details here.

Fields

id          :: Snowflake
unavailable :: Union{Bool, Missing}

source

Discord.VerificationLevel — Type.

A Guild's verification level. More details here.

source

Discord.MessageNotificationLevel — Type.

A Guild's default message notification level. More details here.

source

Discord.ExplicitContentFilterLevel — Type.

A Guild's explicit content filter level. More details here.

source

Discord.MFALevel — Type.

A Guild's MFA level. More details here.

source

Discord.GuildEmbed — Type.

A Guild embed. More details here.

Fields

enabled    :: Bool
channel_id :: Union{Snowflake, Nothing}

source

Discord.Integration — Type.

A Guild integration. More details here.

Fields

id                  :: Snowflake
name                :: String
type                :: String
enabled             :: Bool
syncing             :: Bool
role_id             :: Snowflake
expire_behaviour    :: Int
expire_grace_period :: Int
user                :: User
account             :: IntegrationAccount
synced_at           :: DateTime

source

Discord.IntegrationAccount — Type.

An Integration account. More details here.

Fields

id   :: String
name :: String

source

Discord.Invite — Type.

An invite to a Guild. More details here.

Fields

code                       :: String
guild                      :: Union{Guild, Missing}
channel                    :: DiscordChannel
approximate_presence_cound :: Union{Int, Missing}
approximate_member_count   :: Union{Int, Missing}

source

Discord.InviteMetadata — Type.

Metadata for an Invite. More details here.

Fields

inviter    :: User
uses       :: Int
max_uses   :: Int
max_age    :: Int
temporary  :: Bool
created_at :: DateTime
revoked    :: Bool

source

Discord.Member — Type.

A Guild member. More details here.

Fields

user      :: Union{User, Missing}
nick      :: Union{String, Missing, Nothing}
roles     :: Vector{Snowflake}
joined_at :: DateTime
deaf      :: Bool
mute      :: Bool

source

Discord.Message — Type.

A message sent to a DiscordChannel. More details here.

Fields

id               :: Snowflake
channel_id       :: Snowflake
guild_id         :: Union{Snowflake, Missing}
author           :: Union{User, Missing}
member           :: Union{Member, Missing}
content          :: Union{String, Missing}
timestamp        :: Union{DateTime, Missing}
edited_timestamp :: Union{DateTime, Missing, Nothing}
tts              :: Union{Bool, Missing}
mention_everyone :: Union{Bool, Missing}
mentions         :: Union{Vector{User}, Missing}
mention_roles    :: Union{Vector{Snowflake}, Missing}
attachments      :: Union{Vector{Attachment}, Missing}
embeds           :: Union{Vector{Embed}, Missing}
reactions        :: Union{Vector{Reaction}, Missing}
nonce            :: Union{Snowflake, Missing, Nothing}
pinned           :: Union{Bool, Missing}
webhook_id       :: Union{Snowflake, Missing}
type             :: Union{MessageType, Missing}
activity         :: Union{MessageActivity, Missing}
application      :: Union{MessageApplication, Missing}

source

Discord.MessageActivity — Type.

A Message activity. More details here.

Fields

type     :: MessageActivityType
party_id :: Union{String, Missing}

source

Discord.MessageApplication — Type.

A Rich Presence Message's application information. More details here.

Fields

id          :: Snowflake
cover_image :: String
description :: String
icon        :: String
name        :: String

source

Discord.MessageType — Type.

A Message's type. More details here.

source

Discord.MessageActivityType — Type.

A Message's activity type. More details here.

source

Discord.Overwrite — Type.

A permission overwrite. More details here.

Fields

id    :: Snowflake
type  :: OverwriteType
allow :: Int
deny  :: Int

source

Discord.OverwriteType — Type.

An Overwrite's type. More details here.

source

Discord.Presence — Type.

A User's presence. More details here.

Fields

user       :: User
roles      :: Union{Vector{Snowflake}, Missing}
game       :: Union{Activity, Nothing}
guild_id   :: Union{Snowflake, Missing}
status     :: PresenceStatus
activities :: Vector{Activity}

source

Discord.PresenceStatus — Type.

A User's status sent in a Presence. Available values are PS_IDLE, PS_DND, PS_ONLINE, and PS_OFFLINE. More details here.

source

Discord.Reaction — Type.

A Message reaction. More details here.

Fields

count :: Int
me    :: Bool
emoji :: Emoji

source

Discord.Role — Type.

A User role. More details here.

Fields

id          :: Snowflake
name        :: String
color       :: Union{Int, Missing}
hoist       :: Union{Bool, Missing}
position    :: Union{Int, Missing}
permissions :: Union{Int, Missing}
managed     :: Union{Bool, Missing}
mentionable :: Union{Bool, Missing}

source

Discord.User — Type.

A Discord user. More details here.

Fields

id            :: Snowflake
username      :: Union{String, Missing}
discriminator :: Union{String, Missing}
avatar        :: Union{String, Missing, Nothing}
bot           :: Union{Bool, Missing}
mfa_enabled   :: Union{Bool, Missing}
locale        :: Union{String, Missing}
verified      :: Union{Bool, Missing}
email         :: Union{String, Missing, Nothing}

source

Discord.VoiceRegion — Type.

A region for a Guild's voice server. More details here.

Fields

id         :: String
name       :: String
vip        :: Bool
optimal    :: Bool
deprecated :: Bool
custom     :: Bool

source

Discord.VoiceState — Type.

A User's voice connection status. More details here.

Fields

guild_id   :: Union{Snowflake, Missing}
channel_id :: Union{Snowflake, Nothing}
user_id    :: Snowflake
member     :: Union{Member, Missing}
session_id :: String
deaf       :: Bool
mute       :: Bool
self_deaf  :: Bool
self_mute  :: Bool
suppress   :: Bool

source

Discord.Webhook — Type.

A Webhook. More details here.

Fields

id         :: Snowflake
guild_id   :: Union{Snowflake, Missing}
channel_id :: Snowflake
user       :: Union{User, Missing}
name       :: Union{String, Nothing}
avatar     :: Union{String, Nothing}
token      :: Union{String, Missing}

source