Types

Types

This page is organized in mostly-alphabetical order. Note that Snowflake === UInt64. Unions with Nothing indicate that a field is nullable, whereas Unions with Missing indicate that a field is optional. More details here.

Discord.Activity — Type.

A User activity. More details here.

Fields

name           :: String
type           :: ActivityType
url            :: Union{Missing, Nothing, String}
timestamps     :: Union{Missing, ActivityTimestamps}
application_id :: Union{Missing, UInt64}
details        :: Union{Missing, Nothing, String}
state          :: Union{Missing, Nothing, String}
party          :: Union{Missing, ActivityParty}
assets         :: Union{Missing, ActivityAssets}
secrets        :: Union{Missing, ActivitySecrets}
instance       :: Union{Missing, Bool}
flags          :: Union{Missing, Int64}

source

Discord.ActivityTimestamps — Type.

The start and stop times of an Activity. More details here.

Fields

start :: Union{Missing, DateTime}
stop  :: Union{Missing, DateTime}

source

Discord.ActivityParty — Type.

The current party of an Activity's player. More details here.

Fields

id   :: Union{Missing, String}
size :: Union{Missing, Array{Int64,1}}

source

Discord.ActivityAssets — Type.

Images and hover text for an Activity. More details here.

Fields

large_image :: Union{Missing, String}
large_text  :: Union{Missing, String}
small_image :: Union{Missing, String}
small_text  :: Union{Missing, String}

source

Discord.ActivitySecrets — Type.

Secrets for Rich Presence joining and spectating of an Activity. More details here.

Fields

join     :: Union{Missing, String}
spectate :: Union{Missing, String}
match    :: Union{Missing, String}

source

Discord.ActivityType — Type.

An Activity's type. More details here.

source

Discord.ActivityFlags — Type.

Flags which indicate what an Activity payload contains. More details here.

source

Discord.Attachment — Type.

A Message attachment. More details here.

Fields

id        :: UInt64
filename  :: String
size      :: Int64
url       :: String
proxy_url :: String
height    :: Union{Missing, Int64}
width     :: Union{Missing, Int64}

source

Discord.AuditLog — Type.

An audit log. More details here.

Fields

webhooks          :: Array{Webhook,1}
users             :: Array{User,1}
audit_log_entries :: Array{AuditLogEntry,1}

source

Discord.AuditLogEntry — Type.

An entry in an AuditLog. More details here.

Fields

target_id   :: Union{Nothing, UInt64}
changes     :: Union{Missing, Array{AuditLogChange,1}}
user_id     :: UInt64
id          :: UInt64
action_type :: ActionType
options     :: Union{Missing, AuditLogOptions}
reason      :: Union{Missing, String}

source

Discord.AuditLogChange — Type.

A change item in an AuditLogEntry.

The first type parameter is the type of new_value and old_value. The second is the type of the entity that new_value and old_value belong(ed) to.

More details here.

Fields

new_value :: Union{Missing, T} where T
old_value :: Union{Missing, T} where T
key       :: String
type      :: Type{U} where U

source

Discord.AuditLogOptions — Type.

Optional information in an AuditLogEntry. More details here.

Fields

delete_member_days :: Union{Missing, Int64}
members_removed    :: Union{Missing, Int64}
channel_id         :: Union{Missing, UInt64}
count              :: Union{Missing, Int64}
id                 :: Union{Missing, UInt64}
type               :: Union{Missing, OverwriteType}
role_name          :: Union{Missing, String}

source

Discord.ActionType — Type.

AuditLog action types. More details here.

source

Discord.Ban — Type.

A User ban. More details here.

Fields

reason :: Union{Nothing, String}
user   :: User

source

Discord.DiscordChannel — Type.

A Discord channel. More details here. Note: The name Channel is already used, hence the prefix.

Fields

id                    :: UInt64
type                  :: ChannelType
guild_id              :: Union{Missing, UInt64}
position              :: Union{Missing, Int64}
permission_overwrites :: Union{Missing, Array{Overwrite,1}}
name                  :: Union{Missing, String}
topic                 :: Union{Missing, Nothing, String}
nsfw                  :: Union{Missing, Bool}
last_message_id       :: Union{Missing, Nothing, UInt64}
bitrate               :: Union{Missing, Int64}
user_limit            :: Union{Missing, Int64}
rate_limit_per_user   :: Union{Missing, Int64}
recipients            :: Union{Missing, Array{User,1}}
icon                  :: Union{Missing, Nothing, String}
owner_id              :: Union{Missing, UInt64}
application_id        :: Union{Missing, UInt64}
parent_id             :: Union{Missing, Nothing, UInt64}
last_pin_timestamp    :: Union{Missing, Nothing, DateTime}

source

Discord.ChannelType — Type.

A DiscordChannel's type.

source

Discord.Connection — Type.

A User connection to an external service (Twitch, YouTube, etc.). More details here.

Fields

id           :: String
name         :: String
type         :: String
revoked      :: Bool
integrations :: Array{Integration,1}

source

Discord.Embed — Type.

A Message embed. More details here.

Fields

title       :: Union{Missing, String}
type        :: Union{Missing, String}
description :: Union{Missing, String}
url         :: Union{Missing, String}
timestamp   :: Union{Missing, DateTime}
color       :: Union{Missing, Int64}
footer      :: Union{Missing, EmbedFooter}
image       :: Union{Missing, EmbedImage}
thumbnail   :: Union{Missing, EmbedThumbnail}
video       :: Union{Missing, EmbedVideo}
provider    :: Union{Missing, EmbedProvider}
author      :: Union{Missing, EmbedAuthor}
fields      :: Union{Missing, Array{EmbedField,1}}

source

Discord.EmbedThumbnail — Type.

An Embed's thumbnail image information. More details here.

Fields

url       :: Union{Missing, String}
proxy_url :: Union{Missing, String}
height    :: Union{Missing, Int64}
width     :: Union{Missing, Int64}

source

Discord.EmbedVideo — Type.

An Embed's video information. More details here.

Fields

url    :: Union{Missing, String}
height :: Union{Missing, Int64}
width  :: Union{Missing, Int64}

source

Discord.EmbedImage — Type.

An Embed's image information. More details here.

Fields

url       :: Union{Missing, String}
proxy_url :: Union{Missing, String}
height    :: Union{Missing, Int64}
width     :: Union{Missing, Int64}

source

Discord.EmbedProvider — Type.

An Embed's provider information. More details here.

Fields

name :: Union{Missing, String}
url  :: Union{Missing, Nothing, String}

source

Discord.EmbedAuthor — Type.

An Embed's author information. More details here.

Fields

name           :: Union{Missing, String}
url            :: Union{Missing, String}
icon_url       :: Union{Missing, String}
proxy_icon_url :: Union{Missing, String}

source

Discord.EmbedFooter — Type.

An Embed's footer information. More details here.

Fields

text           :: String
icon_url       :: Union{Missing, String}
proxy_icon_url :: Union{Missing, String}

source

Discord.EmbedField — Type.

An Embed field. More details here.

Fields

name   :: String
value  :: String
inline :: Union{Missing, Bool}

source

Discord.Emoji — Type.

An emoji. More details here.

Fields

id             :: Union{Nothing, UInt64}
name           :: String
roles          :: Union{Missing, Array{UInt64,1}}
user           :: Union{Missing, User}
require_colons :: Union{Missing, Bool}
managed        :: Union{Missing, Bool}
animated       :: Union{Missing, Bool}

source

Discord.AbstractGuild — Type.

A guild (server). Can either be an UnavailableGuild or a Guild.

source

Discord.Guild — Type.

A guild (server). More details here.

Fields

id                            :: UInt64
name                          :: String
icon                          :: Union{Nothing, String}
splash                        :: Union{Nothing, String}
owner                         :: Union{Missing, Bool}
owner_id                      :: Union{Missing, UInt64}
permissions                   :: Union{Missing, Int64}
region                        :: Union{Missing, String}
afk_channel_id                :: Union{Missing, Nothing, UInt64}
afk_timeout                   :: Union{Missing, Int64}
embed_enabled                 :: Union{Missing, Bool}
embed_channel_id              :: Union{Missing, Nothing, UInt64}
verification_level            :: VerificationLevel
default_message_notifications :: Union{Missing, MessageNotificationLevel}
explicit_content_filter       :: Union{Missing, ExplicitContentFilterLevel}
roles                         :: Union{Missing, Array{Role,1}}
emojis                        :: Union{Missing, Array{Emoji,1}}
features                      :: Array{String,1}
mfa_level                     :: Union{Missing, MFALevel}
application_id                :: Union{Missing, Nothing, UInt64}
widget_enabled                :: Union{Missing, Bool}
widget_channel_id             :: Union{Missing, Nothing, UInt64}
system_channel_id             :: Union{Missing, Nothing, UInt64}
joined_at                     :: Union{Missing, DateTime}
large                         :: Union{Missing, Bool}
unavailable                   :: Union{Missing, Bool}
member_count                  :: Union{Missing, Int64}
voice_states                  :: Union{Missing, Array{VoiceState,1}}
members                       :: Union{Missing, Array{Member,1}}
channels                      :: Union{Missing, Array{DiscordChannel,1}}
presences                     :: Union{Missing, Array{Presence,1}}

source

Discord.UnavailableGuild — Type.

An unavailable guild (server). More details here.

Fields

id          :: UInt64
unavailable :: Union{Missing, Bool}

source

Discord.VerificationLevel — Type.

A Guild's verification level. More details here.

source

Discord.MessageNotificationLevel — Type.

A Guild's default message notification level. More details here.

source

Discord.ExplicitContentFilterLevel — Type.

A Guild's explicit content filter level. More details here.

source

Discord.MFALevel — Type.

A Guild's MFA level. More details here.

source

Discord.GuildEmbed — Type.

A Guild embed. More details here.

Fields

enabled    :: Bool
channel_id :: Union{Nothing, UInt64}

source

Discord.Integration — Type.

A Guild integration. More details here.

Fields

id                  :: UInt64
name                :: String
type                :: String
enabled             :: Bool
syncing             :: Bool
role_id             :: UInt64
expire_behaviour    :: Int64
expire_grace_period :: Int64
user                :: User
account             :: IntegrationAccount
synced_at           :: Dates.DateTime

source

Discord.IntegrationAccount — Type.

An Integration account. More details here.

Fields

id   :: String
name :: String

source

Discord.Invite — Type.

An invite to a Guild. More details here.

Fields

code                       :: String
guild                      :: Union{Missing, Guild}
channel                    :: DiscordChannel
approximate_presence_cound :: Union{Missing, Int64}
approximate_member_count   :: Union{Missing, Int64}

source

Discord.InviteMetadata — Type.

Metadata for an Invite. More details here.

Fields

inviter    :: User
uses       :: Int64
max_uses   :: Int64
max_age    :: Int64
temporary  :: Bool
created_at :: Dates.DateTime
revoked    :: Bool

source

Discord.Member — Type.

A Guild member. More details here.

Fields

user      :: Union{Missing, User}
nick      :: Union{Missing, Nothing, String}
roles     :: Array{UInt64,1}
joined_at :: Dates.DateTime
deaf      :: Bool
mute      :: Bool

source

Discord.Message — Type.

A message. More details here.

Fields

id               :: UInt64
channel_id       :: UInt64
guild_id         :: Union{Missing, UInt64}
author           :: Union{Missing, User}
member           :: Union{Missing, Member}
content          :: Union{Missing, String}
timestamp        :: Union{Missing, DateTime}
edited_timestamp :: Union{Missing, Nothing, DateTime}
tts              :: Union{Missing, Bool}
mention_everyone :: Union{Missing, Bool}
mentions         :: Union{Missing, Array{User,1}}
mention_roles    :: Union{Missing, Array{UInt64,1}}
attachments      :: Union{Missing, Array{Attachment,1}}
embeds           :: Union{Missing, Array{Embed,1}}
reactions        :: Union{Missing, Array{Reaction,1}}
nonce            :: Union{Missing, Nothing, UInt64}
pinned           :: Union{Missing, Bool}
webhook_id       :: Union{Missing, UInt64}
type             :: Union{Missing, MessageType}
activity         :: Union{Missing, MessageActivity}
application      :: Union{Missing, MessageApplication}

source

Discord.MessageActivity — Type.

A Message activity. More details here.

Fields

type     :: MessageActivityType
party_id :: Union{Missing, String}

source

Discord.MessageApplication — Type.

A Rich Presence Message's application information. More details here.

Fields

id          :: UInt64
cover_image :: String
description :: String
icon        :: String
name        :: String

source

Discord.MessageType — Type.

A Message's type. More details here.

source

Discord.MessageActivityType — Type.

A Message's activity type. More details here.

source

Discord.Overwrite — Type.

A permission overwrite. More details here.

Fields

id    :: UInt64
type  :: OverwriteType
allow :: Int64
deny  :: Int64

source

Discord.OverwriteType — Type.

An Overwrite's type. More details here.

source

Discord.Presence — Type.

A User's presence. More details here.

Fields

user       :: User
roles      :: Union{Missing, Array{UInt64,1}}
game       :: Union{Nothing, Activity}
guild_id   :: Union{Missing, UInt64}
status     :: PresenceStatus
activities :: Array{Activity,1}

source

Discord.PresenceStatus — Type.

A User's status sent in a Presence. More details here.

source

Discord.Reaction — Type.

A Message reaction. More details here.

Fields

count :: Int64
me    :: Bool
emoji :: Emoji

source

Discord.Role — Type.

A User role. More details here.

Fields

id          :: UInt64
name        :: String
color       :: Int64
hoist       :: Bool
position    :: Int64
permissions :: Int64
managed     :: Bool
mentionable :: Bool

source

Discord.User — Type.

A Discord user. More details here.

Fields

id            :: UInt64
username      :: Union{Missing, String}
discriminator :: Union{Missing, String}
avatar        :: Union{Missing, Nothing, String}
bot           :: Union{Missing, Bool}
mfa_enabled   :: Union{Missing, Bool}
locale        :: Union{Missing, String}
verified      :: Union{Missing, Bool}
email         :: Union{Missing, Nothing, String}

source

Discord.VoiceRegion — Type.

A region for a Guild's voice server. More details here.

Fields

id         :: String
name       :: String
vip        :: Bool
optimal    :: Bool
deprecated :: Bool
custom     :: Bool

source

Discord.VoiceState — Type.

A User's voice connection status. More details here.

Fields

guild_id   :: Union{Missing, UInt64}
channel_id :: Union{Nothing, UInt64}
user_id    :: UInt64
member     :: Union{Missing, Member}
session_id :: String
deaf       :: Bool
mute       :: Bool
self_deaf  :: Bool
self_mute  :: Bool
suppress   :: Bool

source

Discord.Webhook — Type.

A Webhook. More details here.

Fields

id         :: UInt64
guild_id   :: Union{Missing, UInt64}
channel_id :: UInt64
user       :: Union{Missing, User}
name       :: Union{Nothing, String}
avatar     :: Union{Nothing, String}
token      :: String

source